% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiclass-calc-acc-macro.R
\name{calc_acc_macro}
\alias{calc_acc_macro}
\alias{calc_acc_macro.default}
\alias{calc_acc_macro.table}
\alias{calc_acc_macro.data.frame}
\title{Calculate Accuracy Macro.}
\usage{
calc_acc_macro(...)

\method{calc_acc_macro}{default}(tp, tn, n, ...)

\method{calc_acc_macro}{table}(tbl, ...)

\method{calc_acc_macro}{data.frame}(data, prediction, reference, ...)
}
\arguments{
\item{...}{Additional arguments. Not used.}

\item{tp}{Numeric vector of True Positives (TP) by class.}

\item{tn}{Numeric vector of True Negatives (TN) by class.}

\item{n}{Total number of observations.}

\item{tbl}{A table representing the input confusion matrix. This must always have prediction on rows and reference on columns, otherwise most functions in rmetrics will generate incorrect results.}

\item{data}{A data.frame containing the prediction and the reference.}

\item{prediction}{Character. The name of the variable in data that contains the predictions.}

\item{reference}{Character. The name of the variable in data that contains the reference values.}
}
\description{
Calculate Accuracy Macro.
}
\section{Methods (by class)}{
\itemize{
\item \code{calc_acc_macro(default)}: 

\item \code{calc_acc_macro(table)}: 

\item \code{calc_acc_macro(data.frame)}: 

}}
