% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary-calc-af.R
\name{calc_af}
\alias{calc_af}
\title{Calculate Adjusted F-score (af).}
\usage{
calc_af(tp, fp, fn, tn)
}
\arguments{
\item{tp}{Number of true positives in the contingency table.}

\item{fp}{Number of false positives in the contingency table.}

\item{fn}{Number of false negatives in the contingency table.}

\item{tn}{Number of true negatives in the contingency table.}
}
\description{
Calculate Adjusted F-score (af).
}
