% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary-calc-agm.R
\name{calc_agm}
\alias{calc_agm}
\title{Calculate Adjusted geometric mean (agm).}
\usage{
calc_agm(tn, fp, tp, fn)
}
\arguments{
\item{tn}{Number of true negatives in the contingency table.}

\item{fp}{Number of false positives in the contingency table.}

\item{tp}{Number of true positives in the contingency table.}

\item{fn}{Number of false negatives in the contingency table.}
}
\description{
Calculate Adjusted geometric mean (agm).
}
