% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiclass-calc-aickin.R
\name{calc_aickin}
\alias{calc_aickin}
\title{Calculate Aickin's Alpha}
\source{
\enumerate{
\item Aickin, M. (1990) Maximum Likelihood Estimation of Agreement in the Constant Predictive Probability Model, and Its Relation to Cohen's Kappa. Biometrics 46, 293-302.
}
}
\usage{
calc_aickin(
  tbl,
  d = diag(1, nrow = nrow(tbl), ncol = ncol(tbl)),
  epsilon = 1e-07,
  ci.type = "aickin",
  ci.level = 0.95,
  maxiter = 1000
)
}
\arguments{
\item{tbl}{input confusion matrix of the form \code{table(prediction, reference)}}

\item{d}{The agreement matrix for tbl. Must have the same dimensions as tbl, with 1 to indicate agreeing scores, and 0 disagreeing scores. By default the diagonal is considered agreeing.}

\item{epsilon}{Convergence criterion. The algorithm stops when two consecutive \eqn{\alpha} estimates differ by less than epsilon. Default is 1e-7.}

\item{ci.type}{Either FALSE if no confidence intervals are desired or 'aickin' if the default CI's by Aickin should be calculated. If FALSE overwrites ci.level. Default is to calculate the CI.}

\item{ci.level}{A double between 0 and 1 for the levels of the confidence intervals that should be calculated. Default is 0.95.}

\item{maxiter}{Integer or double. Maximum number of iterations to try until convergence. Default is 1000.}
}
\description{
Calculate Aickin's Alpha
}
\author{
Shamelessly stolen from Joseph, L. and Bélisle, P. \url{https://www.medicine.mcgill.ca/epidemiology/Joseph/PBelisle/Aickin-Alpha-Agreement-R.html}
}
