% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiclass-calc-ari.R
\name{calc_ari}
\alias{calc_ari}
\title{Calculate Adjusted Rand index (ari).}
\usage{
calc_ari(tbl, on)
}
\arguments{
\item{tbl}{input confusion matrix of the form \code{table(prediction, reference)}}

\item{on}{Total number of observations}
}
\description{
Calculate Adjusted Rand index (ari).
}
