% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiclass-calc-aunp.R
\name{calc_aunp}
\alias{calc_aunp}
\alias{calc_aunp.data.frame}
\alias{calc_aunp.table}
\title{Calculate overall AUC (aunp).}
\usage{
calc_aunp(...)

\method{calc_aunp}{data.frame}(data, prediction = "prediction", reference = "reference")

\method{calc_aunp}{table}(tbl)
}
\arguments{
\item{data}{data with prediction, reference}

\item{prediction}{Name of the variable in data that holds the predictions}

\item{reference}{Name of the variable in data that holds the reference values}

\item{tbl}{table}
}
\description{
Calculate overall AUC (aunp).
}
\section{Methods (by class)}{
\itemize{
\item \code{calc_aunp(data.frame)}: 

\item \code{calc_aunp(table)}: 

}}
