% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiclass-calc-aunu.R
\name{calc_aunu}
\alias{calc_aunu}
\title{Calculate average AUC (aunu) / AUC macro.}
\usage{
calc_aunu(data, classes)
}
\arguments{
\item{data}{data with prediction, reference}

\item{classes}{confusion matrix classes}
}
\description{
Calculate average AUC (aunu) / AUC macro.
}
