% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary-calc-auroc.R
\name{calc_auroc}
\alias{calc_auroc}
\alias{calc_auroc.default}
\alias{calc_auroc.table}
\alias{calc_auroc.data.frame}
\title{Calculate Area under the ROC curve for each class (auroc).}
\usage{
calc_auroc(...)

\method{calc_auroc}{default}(tn, fp, tp, fn, ...)

\method{calc_auroc}{table}(tbl, ...)

\method{calc_auroc}{data.frame}(data, prediction, reference, ...)
}
\arguments{
\item{...}{Additional arguments. Not used.}

\item{tn}{Numeric, True Negatives (TN).}

\item{fp}{Numeric, False Positives (FP).}

\item{tp}{Numeric, True Positives (TP).}

\item{fn}{Numeric, Fase Negatives (FN).}

\item{tbl}{A table representing the input confusion matrix. This must always have
prediction on rows and reference on columns, otherwise most functions in
rmetrics will generate incorrect results.}

\item{data}{A data.frame containing the prediction and the reference.}

\item{prediction}{Character. The name of the variable in data that contains the predictions.}

\item{reference}{Character. The name of the variable in data that contains the reference
values.}
}
\description{
The area under the curve (AUC) is a metric that is used to evaluate the
performance of a binary classification model. It is calculated by measuring
the area under the curve generated by plotting the true positive rate (TPR)
and false positive rate (FPR) as functions of a probability threshold.
}
\details{
The TPR and FPR are two metrics that are commonly used to evaluate the
performance of a binary classification model. The TPR is the ratio of true
positive predictions to the total number of actual positive samples, while
the FPR is the ratio of false positive predictions to the total number of
actual negative samples.

The AUC can be calculated by first generating a series of TPR-FPR pairs at
different probability thresholds. This is done by varying the threshold used
to classify a sample as positive or negative, and calculating the TPR and
FPR at each threshold. These pairs are then plotted on a graph, with the
TPR on the y-axis and the FPR on the x-axis. The AUC is then calculated
by measuring the area under this curve.

The AUC is a useful metric because it allows for a more comprehensive
evaluation of a model's performance. The AUC takes into account the entire
TPR-FPR curve and thereby provides a more complete picture of a model's
performance.
}
\section{Methods (by class)}{
\itemize{
\item \code{calc_auroc(default)}: 

\item \code{calc_auroc(table)}: 

\item \code{calc_auroc(data.frame)}: 

}}
\seealso{
\code{\link[=calc_tpr]{calc_tpr()}}

\code{\link[=calc_fpr]{calc_fpr()}}

\code{\link[=calc_aupr]{calc_aupr()}}
}
