% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiclass-calc-b.R
\name{calc_b}
\alias{calc_b}
\title{Calculate Bangdiwala's B.}
\usage{
calc_b(tp, ppos, pos)
}
\arguments{
\item{tp}{vector of true positives per class}

\item{ppos}{vector of positives in predict vector per class}

\item{pos}{vector of actual positives per class}
}
\description{
Calculate Bangdiwala's B.
}
