% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary-calc-bbs.R
\name{calc_bbs}
\alias{calc_bbs}
\title{Calculate Braun-Blanquet similarity (bbs).}
\usage{
calc_bbs(tp, ppos, pos)
}
\arguments{
\item{tp}{Number of true positives in the contingency table.}

\item{ppos}{Number of positives in predict vector}

\item{pos}{Number of actual positives}
}
\description{
Calculate Braun-Blanquet similarity (bbs).
}
