% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiclass-calc-cba.R
\name{calc_cba}
\alias{calc_cba}
\title{Calculate Class balance accuracy (cba).}
\usage{
calc_cba(tbl, ppos, pos)
}
\arguments{
\item{tbl}{input confusion matrix of the form \code{table(prediction, reference)}}

\item{ppos}{vector of positives in prediction by class}

\item{pos}{vector of positives in reference by class}
}
\description{
Calculate Class balance accuracy (cba).
}
