% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiclass-calc-conditional-entropy.R
\name{calc_conditional_entropy}
\alias{calc_conditional_entropy}
\alias{calc_conditional_entropy.data.frame}
\alias{calc_conditional_entropy.table}
\title{Calculate conditional entropy.}
\usage{
calc_conditional_entropy(...)

\method{calc_conditional_entropy}{data.frame}(
  data,
  prediction = "prediction",
  reference = "reference"
)

\method{calc_conditional_entropy}{table}(tbl)
}
\arguments{
\item{data}{data.frame}

\item{tbl}{confusion matrix}
}
\description{
Computes the average conditional entropy between two vectors.
}
\section{Methods (by class)}{
\itemize{
\item \code{calc_conditional_entropy(data.frame)}: 

\item \code{calc_conditional_entropy(table)}: 

}}
