% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiclass-calc-csi-macro.R
\name{calc_csi_macro}
\alias{calc_csi_macro}
\title{Calculate classification success index macro (csi macro).}
\usage{
calc_csi_macro(tp, fn, fp)
}
\arguments{
\item{tp}{TP by class}

\item{fn}{FN by class}

\item{fp}{FP by class}
}
\description{
Calculate classification success index macro (csi macro).
}
