% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary-calc-dind.R
\name{calc_dind}
\alias{calc_dind}
\title{Calculate Distance index (dind).}
\usage{
calc_dind(tn, fp, tp, fn)
}
\arguments{
\item{tn}{Number of true negatives in the contingency table.}

\item{fp}{Number of false positives in the contingency table.}

\item{tp}{Number of true positives in the contingency table.}

\item{fn}{Number of false negatives in the contingency table.}
}
\description{
Calculate Distance index (dind).
}
