% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary-calc-dp.R
\name{calc_dp}
\alias{calc_dp}
\alias{calc_dp.default}
\alias{calc_dp.table}
\alias{calc_dp.data.frame}
\title{Calculate Discriminant Power.}
\usage{
calc_dp(...)

\method{calc_dp}{default}(tn, fp, tp, fn, ...)

\method{calc_dp}{table}(tbl, ...)

\method{calc_dp}{data.frame}(data, prediction, reference, ...)
}
\arguments{
\item{...}{Additional arguments. Not used.}

\item{tn}{Numeric, True Negatives (TN).}

\item{fp}{Numeric, False Positives (FP).}

\item{tp}{Numeric, True Positives (TP).}

\item{fn}{Numeric, Fase Negatives (FN).}

\item{tbl}{A table representing the input confusion matrix. This must always have
prediction on rows and reference on columns, otherwise most functions in
rmetrics will generate incorrect results.}

\item{data}{A data.frame containing the prediction and the reference.}

\item{prediction}{Character. The name of the variable in data that contains the predictions.}

\item{reference}{Character. The name of the variable in data that contains the reference
values.}
}
\description{
Calculate Discriminant Power.
}
\section{Methods (by class)}{
\itemize{
\item \code{calc_dp(default)}: 

\item \code{calc_dp(table)}: 

\item \code{calc_dp(data.frame)}: 

}}
