% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary-calc-f.R
\name{calc_f}
\alias{calc_f}
\alias{calc_f.default}
\alias{calc_f.table}
\alias{calc_f.data.frame}
\title{Calculate (adjusted) F-Score}
\usage{
calc_f(...)

\method{calc_f}{default}(tp, fp, fn, beta, ...)

\method{calc_f}{table}(tbl, beta, incr = FALSE, ...)

\method{calc_f}{data.frame}(data, prediction, reference, beta, incr = FALSE, ...)
}
\arguments{
\item{...}{Additional arguments. Not used.}

\item{tp}{Numeric, True Positives (TP).}

\item{fp}{Numeric, False Positives (FP).}

\item{fn}{Numeric, Fase Negatives (FN).}

\item{beta}{beta coefficient}

\item{tbl}{A table representing the input confusion matrix. This must always have
prediction on rows and reference on columns, otherwise most functions in
rmetrics will generate incorrect results.}

\item{incr}{Double. Continuity correction to add to each cell of the table. FALSE
by default, which means the raw counts will be used.}

\item{data}{A data.frame containing the prediction and the reference.}

\item{prediction}{Character. The name of the variable in data that contains the predictions.}

\item{reference}{Character. The name of the variable in data that contains the reference
values.}
}
\description{
The F-score, also known as the F1-score or F-measure, is a metric that
combines precision and recall into a single score. This implementation
allows you to set the parameter \code{beta} to control the relative weight of precision and recall.
}
\details{
To calculate the adjusted F-score, the following formula is used:

Adjusted F-score = (1 + beta^2) \* (Precision \* Recall) / (beta^2 \* Precision + Recall)

where beta is a parameter that controls the relative weight of precision
and recall. For beta = 1, the adjusted F-score is equivalent to the F1-score.
For beta < 1, precision is given more weight, and for beta > 1, recall is
given more weight.

Overall, the adjusted F-score is a useful metric for evaluating the
performance of a predictive model, particularly when the precision and
recall are calculated using a sample of the data. It allows for a more
balanced evaluation of a model's performance by taking into account the
number of data points used to calculate the precision and recall.
}
\section{Methods (by class)}{
\itemize{
\item \code{calc_f(default)}: 

\item \code{calc_f(table)}: 

\item \code{calc_f(data.frame)}: 

}}
