% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary-calc-f05.R
\name{calc_f05}
\alias{calc_f05}
\title{Calculate F0.5 Score}
\usage{
calc_f05(tp, fp, fn)
}
\arguments{
\item{tp}{Number of true positives in the contingency table.}

\item{fp}{Number of false positives in the contingency table.}

\item{fn}{Number of false negatives in the contingency table.}
}
\description{
Calculate F0.5 Score
}
