% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary-calc-f1.R
\name{calc_f1}
\alias{calc_f1}
\title{Calculate F1 Score}
\usage{
calc_f1(tp, fp, fn)
}
\arguments{
\item{tp}{Number of true positives in the contingency table.}

\item{fp}{Number of false positives in the contingency table.}

\item{fn}{Number of false negatives in the contingency table.}
}
\description{
Calculate F1 Score
}
