% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiclass-calc-f1-macro.R
\name{calc_f1_macro}
\alias{calc_f1_macro}
\title{Calculate F1 macro.}
\usage{
calc_f1_macro(tp, fp, fn)
}
\arguments{
\item{tp}{TP by class}

\item{fp}{FP by class}

\item{fn}{FN by class}
}
\description{
Calculate F1 macro.
}
