% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary-calc-f2.R
\name{calc_f2}
\alias{calc_f2}
\title{Calculate F2 Score}
\usage{
calc_f2(tp, fp, fn)
}
\arguments{
\item{tp}{Number of true positives in the contingency table.}

\item{fp}{Number of false positives in the contingency table.}

\item{fn}{Number of false negatives in the contingency table.}
}
\description{
Calculate F2 Score
}
