% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiclass-calc-f-macro.R
\name{calc_f_macro}
\alias{calc_f_macro}
\alias{calc_f_macro.default}
\alias{calc_f_macro.table}
\alias{calc_f_macro.data.frame}
\title{Calculate F macro.}
\usage{
calc_f_macro(...)

\method{calc_f_macro}{default}(tp, fp, fn, beta = 1)

\method{calc_f_macro}{table}(tbl)

\method{calc_f_macro}{data.frame}(data, prediction, reference)
}
\arguments{
\item{tp}{Numeric vector of True Positives (TP) by class.}

\item{fp}{Numeric vector of False Positives (FP) by class.}

\item{fn}{Numeric vector of Fase Negatives (FN) by class.}

\item{beta}{Scaling factor. 1 by default for the F1-Score.}

\item{tbl}{A table representing the input confusion matrix. This must always have prediction on rows and reference on columns, otherwise most functions in rmetrics will generate incorrect results.}

\item{data}{A data.frame containing the prediction and the reference.}

\item{prediction}{Character. The name of the variable in data that contains the predictions.}

\item{reference}{Character. The name of the variable in data that contains the reference values.}
}
\description{
Calculate F macro.
}
\section{Methods (by class)}{
\itemize{
\item \code{calc_f_macro(default)}: 

\item \code{calc_f_macro(table)}: 

\item \code{calc_f_macro(data.frame)}: 

}}
