% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary-calc-fmi.R
\name{calc_fmi}
\alias{calc_fmi}
\title{Calculate Fowlkes–Mallows index}
\usage{
calc_fmi(tp, fp, fn)
}
\arguments{
\item{tp}{Number of true positives in the contingency table.}

\item{fp}{Number of false positives in the contingency table.}

\item{fn}{Number of false negatives in the contingency table.}
}
\description{
Calculate Fowlkes–Mallows index
}
