% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiclass-calc-fnr-micro.R
\name{calc_fnr_micro}
\alias{calc_fnr_micro}
\title{Calculate FNR macro.}
\usage{
calc_fnr_micro(otp, on)
}
\arguments{
\item{otp}{Overall true positives}

\item{on}{Overall N}
}
\description{
Calculate FNR macro.
}
