% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary-calc-fpr.R
\name{calc_fpr}
\alias{calc_fpr}
\title{Calculate False Positive Rate}
\usage{
calc_fpr(fp, tn, ci.type, ci.level)
}
\arguments{
\item{fp}{Number of false positives in the contingency table.}

\item{tn}{Number of true negatives in the contingency table.}

\item{ci.type}{Either FALSE if no confidence intervals are desired or one of "agresti.coull", "agresti-coull", "ac", "asymptotic", "normal", "wald", "clopper-pearson", "cp", "exact", "jeffreys", "bayes", and "wilson". If FALSE, overwrites ci.level.}

\item{ci.level}{A number between 0 and 1 for the levels of the confidence intervals that should be calculated.}
}
\description{
Calculate False Positive Rate
}
