% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiclass-calc-fpr-macro.R
\name{calc_fpr_macro}
\alias{calc_fpr_macro}
\title{Calculate FPR macro.}
\usage{
calc_fpr_macro(fp, tn)
}
\arguments{
\item{fp}{FP by class}

\item{tn}{TN by class}
}
\description{
Calculate FPR macro.
}
