% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiclass-calc-fpr-micro.R
\name{calc_fpr_micro}
\alias{calc_fpr_micro}
\title{Calculate FPR macro.}
\usage{
calc_fpr_micro(tn, fp)
}
\arguments{
\item{tn}{TN by class}

\item{fp}{FP by class}
}
\description{
Calculate FPR macro.
}
