% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary-calc-gmean.R
\name{calc_gmean}
\alias{calc_gmean}
\alias{calc_gmean.default}
\alias{calc_gmean.table}
\alias{calc_gmean.data.frame}
\title{Calculate Geometric mean (gmean) of TPR and TNR.}
\usage{
calc_gmean(...)

\method{calc_gmean}{default}(tn, fp, tp, fn, ...)

\method{calc_gmean}{table}(tbl, ...)

\method{calc_gmean}{data.frame}(data, prediction, reference, ...)
}
\arguments{
\item{...}{Additional arguments. Not used.}

\item{tn}{Numeric, True Negatives (TN).}

\item{fp}{Numeric, False Positives (FP).}

\item{tp}{Numeric, True Positives (TP).}

\item{fn}{Numeric, Fase Negatives (FN).}

\item{tbl}{A table representing the input confusion matrix. This must always have
prediction on rows and reference on columns, otherwise most functions in
rmetrics will generate incorrect results.}

\item{data}{A data.frame containing the prediction and the reference.}

\item{prediction}{Character. The name of the variable in data that contains the predictions.}

\item{reference}{Character. The name of the variable in data that contains the reference
values.}
}
\description{
Calculate Geometric mean (gmean) of TPR and TNR.
}
\section{Methods (by class)}{
\itemize{
\item \code{calc_gmean(default)}: 

\item \code{calc_gmean(table)}: 

\item \code{calc_gmean(data.frame)}: 

}}
