% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dual-cutoff-calc-grey.R
\name{calc_grey}
\alias{calc_grey}
\title{Calculate Size of Grey Area}
\usage{
calc_grey(ind, tot, ci.type, ci.level)
}
\arguments{
\item{ind}{Number of indeterminate results.}

\item{tot}{Total number of trials.}

\item{ci.type}{Either FALSE if no confidence intervals are desired or one of 'boot', ADD OTHERS. If FALSE overwrites ci.level and boot.}

\item{ci.level}{A number between 0 and 1 for the levels of the confidence intervals that should be calculated.}
}
\description{
Calculate Size of Grey Area
}
