% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiclass-calc-gwet-ac1.R
\name{calc_gwet_ac1}
\alias{calc_gwet_ac1}
\title{Calculate Gwet AC1.}
\usage{
calc_gwet_ac1(pos, ppos, otp, on)
}
\arguments{
\item{pos}{Vector of positives in reference.}

\item{ppos}{Vector of positives in prediction.}

\item{otp}{Overall true positives.}

\item{on}{Number of samples.}
}
\description{
Calculate Gwet AC1.
}
