% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiclass-calc-hamming.R
\name{calc_hamming}
\alias{calc_hamming}
\title{Calculate Hamming loss.}
\usage{
calc_hamming(otp, on)
}
\arguments{
\item{otp}{Overall true positives}

\item{on}{Onverall N}
}
\description{
Calculate Hamming loss.
}
