% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary-calc-iba.R
\name{calc_iba}
\alias{calc_iba}
\title{Calculate Index of balanced acc (iba).}
\usage{
calc_iba(tn, fp, tp, fn, alpha = 1)
}
\arguments{
\item{tn}{Number of true negatives in the contingency table.}

\item{fp}{Number of false positives in the contingency table.}

\item{tp}{Number of true positives in the contingency table.}

\item{fn}{Number of false negatives in the contingency table.}

\item{alpha}{How much should TPR-TNR be weighted? Default is 1.}
}
\description{
Calculate Index of balanced acc (iba).
}
