% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary-calc-icsi.R
\name{calc_icsi}
\alias{calc_icsi}
\title{Calculate Individual classification success index (icsi).}
\usage{
calc_icsi(tp, fn, fp)
}
\arguments{
\item{tp}{TP}

\item{fn}{FN}

\item{fp}{FP}
}
\description{
Calculate Individual classification success index (icsi).
}
