% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary-calc-informedness.R
\name{calc_informedness}
\alias{calc_informedness}
\title{Calculate Informedness}
\usage{
calc_informedness(tp, fn, tn, fp)
}
\arguments{
\item{tp}{Number of true positives in the contingency table.}

\item{fn}{Number of false negatives in the contingency table.}

\item{tn}{Number of true negatives in the contingency table.}

\item{fp}{Number of false positives in the contingency table.}
}
\description{
Calculate Informedness
}
