% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary-calc-is.R
\name{calc_is}
\alias{calc_is}
\title{Calculate Information score (is).}
\usage{
calc_is(tp, fp, fn, n)
}
\arguments{
\item{tp}{Number of true positives in the contingency table.}

\item{fp}{Number of false positives in the contingency table.}

\item{fn}{Number of false negatives in the contingency table.}

\item{n}{Total number of samples.}
}
\description{
Calculate Information score (is).
}
