% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary-calc-ji.R
\name{calc_ji}
\alias{calc_ji}
\title{Calculate Jaccard index (ji).}
\usage{
calc_ji(tp, ppos, pos)
}
\arguments{
\item{tp}{Number of true positives in the contingency table.}

\item{ppos}{Number of positives in predict vector}

\item{pos}{Number of actual positives}
}
\description{
Calculate Jaccard index (ji).
}
