% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiclass-calc-kl-divergence.R
\name{calc_kl_divergence}
\alias{calc_kl_divergence}
\title{Calculate Kullback-Liebler (KL) divergence.}
\usage{
calc_kl_divergence(pos, ppos, n)
}
\arguments{
\item{pos}{vector of actual positives per class}

\item{ppos}{vector of positives in predict vector per class}

\item{n}{Total number of observations}
}
\description{
Calculate Kullback-Liebler (KL) divergence.
}
