% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiclass-calc-lambda.R
\name{calc_lambda}
\alias{calc_lambda}
\alias{calc_lambda.table}
\alias{calc_lambda.data.frame}
\title{Calculate Goodman Kruskal Lambda.}
\usage{
calc_lambda(...)

\method{calc_lambda}{table}(
  tbl,
  direction = c("symmetric", "row", "column"),
  ci.type = FALSE,
  ci.level = 0.95,
  ...
)

\method{calc_lambda}{data.frame}(
  data,
  prediction = "prediction",
  reference = "reference",
  direction = c("symmetric", "row", "column"),
  ci.type = FALSE,
  ci.level = 0.95,
  ...
)
}
\arguments{
\item{...}{Additional arguments. Not used.

Calculates symmetric and asymmetric (lambda A and lambda B) Goodman Kruskal lambda and their confidence intervals.
Lambda measures the proportional reduction in error in cross tabulation analysis. It can be used the gauge the strength of association between two nominal variables.
It can be interpreted as the probable improvement in predicting the reference given knowledge of the predictions.}

\item{tbl}{A table representing the input confusion matrix. This must always have prediction on rows and reference on columns, otherwise most functions in rmetrics will generate incorrect results.}

\item{direction}{Character, either "symmetric", "row" or "column". "row" corresponds to Lambda B and "column" to Lambda A.}

\item{ci.type}{FALSE if no ci is requested or "normal" for normal approximation CIs}

\item{ci.level}{A number between 0 and 1 for the levels of the confidence intervals that should be calculated.}

\item{data}{A data.frame containing the prediction and the reference.}

\item{prediction}{Character. The name of the variable in data that contains the predictions.}

\item{reference}{Character. The name of the variable in data that contains the reference values.}
}
\value{
A numeric vector with the three elements 'lambda', 'll' and 'ul'. If no CI is requested, 'll' and 'ul' are NA.
}
\description{
Calculate Goodman Kruskal Lambda.
}
\details{
This implementation is based on code by Andri Signorell \href{mailto:andri@signorell.net}{andri@signorell.net}, Antti Arppe \href{mailto:antti.arppe@helsinki.fi}{antti.arppe@helsinki.fi} and Nanina Anderegg (confidence interval symmetric lambda) for the package \code{DescTools}.
}
\section{Methods (by class)}{
\itemize{
\item \code{calc_lambda(table)}: 

\item \code{calc_lambda(data.frame)}: 

}}
\references{
Agresti, A. (2002) Categorical Data Analysis. John Wiley & Sons
Goodman, L. A., Kruskal W. H. (1979) Measures of Association for Cross Classifications. New York: Springer-Verlag (contains articles appearing in J. Amer. Statist. Assoc. in 1954, 1959, 1963, 1972).
Liebetrau, A. M. (1983) Measures of Association, Sage University Papers Series on Quantitative Applications in the Social Sciences, 07-004. Newbury Park, CA: Sage, pp. 17–24
}
