% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary-calc-mcc.R
\name{calc_mcc}
\alias{calc_mcc}
\title{Calculate Matthews correlation coefficient (mcc).}
\usage{
calc_mcc(tp, tn, fp, fn)
}
\arguments{
\item{tp}{Number of true positives in the contingency table.}

\item{tn}{Number of true negatives in the contingency table.}

\item{fp}{Number of false positives in the contingency table.}

\item{fn}{Number of false negatives in the contingency table.}
}
\description{
Calculate Matthews correlation coefficient (mcc).
}
