% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiclass-calc-mcc-overall.R
\name{calc_mcc_overall}
\alias{calc_mcc_overall}
\title{Calculate Calculate Overall_MCC.}
\usage{
calc_mcc_overall(tbl, ppos, pos)
}
\arguments{
\item{tbl}{Classification matrix.}

\item{ppos}{Positives in prediction.}

\item{pos}{Positives in reference.}
}
\description{
Calculate Calculate Overall_MCC.
}
