% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiclass-calc-mutual-information.R
\name{calc_mutual_information}
\alias{calc_mutual_information}
\title{Calculate Mutual information.}
\usage{
calc_mutual_information(tbl, pos, ppos, n)
}
\arguments{
\item{tbl}{confusion matrix}

\item{pos}{vector of actual positives per class}

\item{ppos}{vector of positives in predict vector per class}

\item{n}{Total number of observations}
}
\description{
Calculate Mutual information.
}
