% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary-calc-net-benefit.R
\name{calc_net_benefit}
\alias{calc_net_benefit}
\title{Calculate Net Benefit}
\usage{
calc_net_benefit(tp, fp, n, w)
}
\arguments{
\item{tp}{Number of true positives in the contingency table.}

\item{fp}{Number of false positives in the contingency table.}

\item{n}{Total number of samples}

\item{w}{weight}
}
\description{
Calculate Net Benefit
}
