% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiclass-calc-nir.R
\name{calc_nir}
\alias{calc_nir}
\title{Calculate No information rate (nir).}
\usage{
calc_nir(pos, on)
}
\arguments{
\item{pos}{Positives in reference.}

\item{on}{Overall N}
}
\description{
Calculate No information rate (nir).
}
