% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary-calc-npv.R
\name{calc_npv}
\alias{calc_npv}
\alias{calc_npv.default}
\alias{calc_npv.table}
\alias{calc_npv.data.frame}
\title{Calculate Negative Predictive Value}
\usage{
calc_npv(...)

\method{calc_npv}{default}(tn, fn, ci.type, ci.level, ...)

\method{calc_npv}{table}(tbl, ci.type, ci.level, ...)

\method{calc_npv}{data.frame}(data, prediction, reference, ci.type, ci.level, ...)
}
\arguments{
\item{...}{Additional arguments. Not used.}

\item{tn}{Numeric, True Negatives (TN).}

\item{fn}{Numeric, Fase Negatives (FN).}

\item{ci.type}{Either FALSE if no confidence intervals are desired or one of "agresti.coull", "agresti-coull", "ac", "asymptotic", "normal", "wald", "clopper-pearson", "cp", "exact", "jeffreys", "bayes", and "wilson". If FALSE, overwrites ci.level.}

\item{ci.level}{A number between 0 and 1 for the levels of the confidence intervals that should be calculated.}

\item{tbl}{A table representing the input confusion matrix. This must always have prediction on rows and reference on columns, otherwise most functions in rmetrics will generate incorrect results.}

\item{data}{A data.frame containing the prediction and the reference.}

\item{prediction}{Character. The name of the variable in data that contains the predictions.}

\item{reference}{Character. The name of the variable in data that contains the reference values.}
}
\description{
Calculate Negative Predictive Value
}
\section{Methods (by class)}{
\itemize{
\item \code{calc_npv(default)}: 

\item \code{calc_npv(table)}: 

\item \code{calc_npv(data.frame)}: 

}}
