% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiclass-calc-oacc.R
\name{calc_oacc}
\alias{calc_oacc}
\title{Calculate Calculate Overall accuracy (oacc).}
\usage{
calc_oacc(otp, on)
}
\arguments{
\item{otp}{Overall number of true positives in the contingency table.}

\item{on}{Overall number of observations}
}
\description{
Also called TPR Micro.
}
