% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary-calc-ooc.R
\name{calc_ooc}
\alias{calc_ooc}
\title{Calculate Otsuka-Ochiai coefficient (ooc)}
\usage{
calc_ooc(tp, ppos, pos)
}
\arguments{
\item{tp}{Number of true positives in the contingency table.}

\item{ppos}{Number of positives in predict vector}

\item{pos}{Number of actual positives}
}
\description{
Calculate Otsuka-Ochiai coefficient (ooc)
}
