% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary-calc-op.R
\name{calc_op}
\alias{calc_op}
\title{Calculate Optimized precision (op).}
\usage{
calc_op(tn, fp, tp, fn)
}
\arguments{
\item{tn}{Number of true negatives in the contingency table.}

\item{fp}{Number of false positives in the contingency table.}

\item{tp}{Number of true positives in the contingency table.}

\item{fn}{Number of false negatives in the contingency table.}
}
\description{
Calculate Optimized precision (op).
}
