% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiclass-calc-pearson-c.R
\name{calc_pearson_c}
\alias{calc_pearson_c}
\title{Calculate Pearson's C.}
\usage{
calc_pearson_c(tbl, on)
}
\arguments{
\item{tbl}{confusion matrix}

\item{on}{Total number of observations}
}
\description{
Calculate Pearson's C.
}
