% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary-calc-phi.R
\name{calc_phi}
\alias{calc_phi}
\title{Calculate Phi Coefficient}
\usage{
calc_phi(tp, tn, fp, fn)
}
\arguments{
\item{tp}{Number of true positives in the contingency table.}

\item{tn}{Number of true negatives in the contingency table.}

\item{fp}{Number of false positives in the contingency table.}

\item{fn}{Number of false negatives in the contingency table.}
}
\description{
Calculate Phi Coefficient
}
