% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiclass-calc-phi-square.R
\name{calc_phi_square}
\alias{calc_phi_square}
\title{Calculate Phi-squared.}
\usage{
calc_phi_square(tbl, on)
}
\arguments{
\item{tbl}{confusion matrix}

\item{on}{Total number of observations}
}
\description{
Calculate Phi-squared.
}
