% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiclass-calc-phi-square.R
\name{calc_phisq}
\alias{calc_phisq}
\alias{calc_phisq.table}
\alias{calc_phisq.data.frame}
\title{Calculate Phi-squared.}
\usage{
calc_phisq(...)

\method{calc_phisq}{table}(tbl, ...)

\method{calc_phisq}{data.frame}(data, prediction, reference, ...)
}
\arguments{
\item{...}{Additional arguments passed on to \code{stats::chisq.test}.}

\item{tbl}{A table representing the input confusion matrix. This must always have
prediction on rows and reference on columns, otherwise most functions in
rmetrics will generate incorrect results.}

\item{data}{A data.frame containing the prediction and the reference.}

\item{prediction}{Character. The name of the variable in data that contains the predictions.}

\item{reference}{Character. The name of the variable in data that contains the reference
values.}
}
\description{
Calculate Phi-squared.
}
\section{Methods (by class)}{
\itemize{
\item \code{calc_phisq(table)}: 

\item \code{calc_phisq(data.frame)}: 

}}
