% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary-calc-plr.R
\name{calc_plr}
\alias{calc_plr}
\title{Calculate Positive Likelihood Ratio (plr)}
\source{
Koopman, P. A. R. Confidence intervals for the ratio of two binomial proportions. Biometrics (1984): 513-517.
}
\usage{
calc_plr(tp, fn, fp, tn, ci.type, ci.level)
}
\arguments{
\item{tp}{Number of true positives in the contingency table.}

\item{fn}{Number of false negatives in the contingency table.}

\item{fp}{Number of false positives in the contingency table.}

\item{tn}{Number of true negatives in the contingency table.}

\item{ci.type}{Either FALSE if no confidence intervals are desired or 'koopman'. If FALSE overwrites ci.level and boot.}

\item{ci.level}{A number between 0 and 1 for the levels of the confidence intervals that should be calculated.}
}
\description{
Calculate Positive Likelihood Ratio (plr)
}
