% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiclass-calc-ppv-macro.R
\name{calc_ppv_macro}
\alias{calc_ppv_macro}
\title{Calculate Precision macro (ppv macro).}
\usage{
calc_ppv_macro(tp, fp)
}
\arguments{
\item{tp}{TP by class}

\item{fp}{FP by class}
}
\description{
Calculate Precision macro (ppv macro).
}
