% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary-calc-precision.R
\name{calc_precision}
\alias{calc_precision}
\title{Calculate Precision}
\usage{
calc_precision(tp, fp, ci.type, ci.level)
}
\arguments{
\item{tp}{Number of true positives in the contingency table.}

\item{fp}{Number of false positives in the contingency table.}

\item{ci.type}{Either FALSE if no confidence intervals are desired or one of "agresti.coull", "agresti-coull", "ac", "asymptotic", "normal", "wald", "clopper-pearson", "cp", "exact", "jeffreys", "bayes", and "wilson". If FALSE, overwrites ci.level.}

\item{ci.level}{A number between 0 and 1 for the levels of the confidence intervals that should be calculated.}
}
\description{
Calculate Precision
}
