% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary-calc-prevalence.R
\name{calc_prevalence}
\alias{calc_prevalence}
\title{Calculate Sample Prevalence}
\usage{
calc_prevalence(pos, neg, ci.type, ci.level)
}
\arguments{
\item{pos}{Number of positives according to reference.}

\item{neg}{Number of negatives according to reference.}

\item{ci.type}{Either FALSE if no confidence intervals are desired or one of "agresti.coull", "agresti-coull", "ac", "asymptotic", "normal", "wald", "clopper-pearson", "cp", "exact", "jeffreys", "bayes", and "wilson". If FALSE, overwrites ci.level.}

\item{ci.level}{A number between 0 and 1 for the levels of the confidence intervals that should be calculated.}
}
\description{
Calculate Sample Prevalence
}
