% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary-calc-q.R
\name{calc_q}
\alias{calc_q}
\title{Calculate Yule's Q. .}
\usage{
calc_q(tp, tn, fp, fn)
}
\arguments{
\item{tp}{Number of true positives in the contingency table.}

\item{tn}{Number of true negatives in the contingency table.}

\item{fp}{Number of false positives in the contingency table.}

\item{fn}{Number of false negatives in the contingency table.}
}
\description{
Calculate Yule's Q. .
}
