% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary-calc-racc.R
\name{calc_racc}
\alias{calc_racc}
\title{Calculate Random accuracy}
\usage{
calc_racc(ppos, pos, n)
}
\arguments{
\item{ppos}{Number of positives in predict vector}

\item{pos}{Number of positives in reference vector}

\item{n}{Total number of samples}
}
\description{
Calculate Random accuracy
}
