% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary-calc-raccu.R
\name{calc_raccu}
\alias{calc_raccu}
\title{Calculate Random acc Unbiased}
\usage{
calc_raccu(ppos, pos, n)
}
\arguments{
\item{ppos}{Number of positives in predict vector}

\item{pos}{Number of positives in reference vector}

\item{n}{Total number of samples}
}
\description{
Calculate Random acc Unbiased
}
