% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiclass-calc-rand.R
\name{calc_rand}
\alias{calc_rand}
\title{Calculate Rand index (rand).}
\usage{
calc_rand(prediction, reference)
}
\arguments{
\item{prediction}{prediction}

\item{reference}{reference}
}
\description{
Calculate Rand index (rand).
}
