% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiclass-calc-reference-entropy.R
\name{calc_reference_entropy}
\alias{calc_reference_entropy}
\title{Calculate Reference Entropy.}
\usage{
calc_reference_entropy(pos, n)
}
\arguments{
\item{pos}{vector of actual positives per class}

\item{n}{Total number of observations}
}
\description{
Calculate Reference Entropy.
}
