% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiclass-calc-rr.R
\name{calc_rr}
\alias{calc_rr}
\title{Calculate Global performance index (rr).}
\usage{
calc_rr(ppos)
}
\arguments{
\item{ppos}{Positives in prediction.}
}
\description{
Calculate Global performance index (rr).
}
