% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary-calc-sind.R
\name{calc_sind}
\alias{calc_sind}
\title{Calculate Similarity index (sind).}
\usage{
calc_sind(tn, fp, tp, fn)
}
\arguments{
\item{tn}{Number of true negatives in the contingency table.}

\item{fp}{Number of false positives in the contingency table.}

\item{tp}{Number of true positives in the contingency table.}

\item{fn}{Number of false negatives in the contingency table.}
}
\description{
Calculate Similarity index (sind).
}
