% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiclass-calc-tnr-macro.R
\name{calc_tnr_macro}
\alias{calc_tnr_macro}
\title{Calculate TNR macro.}
\usage{
calc_tnr_macro(tn, fp)
}
\arguments{
\item{tn}{TN by class}

\item{fp}{FP by class}
}
\description{
Calculate TNR macro.
}
