% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiclass-calc-tnr-micro.R
\name{calc_tnr_micro}
\alias{calc_tnr_micro}
\title{Calculate TNR micro.}
\usage{
calc_tnr_micro(tn, fp)
}
\arguments{
\item{tn}{TN by class}

\item{fp}{FP by class}
}
\description{
Calculate TNR micro.
}
