% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary-calc-tpr.R
\name{calc_tpr}
\alias{calc_tpr}
\title{Calculate Sensitivity (tpr)}
\usage{
calc_tpr(tp, fn, ci.type, ci.level)
}
\arguments{
\item{tp}{Number of true positives in the contingency table.}

\item{fn}{Number of false negatives in the contingency table.}

\item{ci.type}{Either FALSE if no confidence intervals are desired or one of "agresti.coull", "agresti-coull", "ac", "asymptotic", "normal", "wald", "clopper-pearson", "cp", "exact", "jeffreys", "bayes", and "wilson". If FALSE, overwrites ci.level.}

\item{ci.level}{A number between 0 and 1 for the levels of the confidence intervals that should be calculated.}
}
\description{
Calculate Sensitivity (tpr)
}
