% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiclass-calc-zero-one-loss.R
\name{calc_zero_one_loss}
\alias{calc_zero_one_loss}
\title{Calculate Zero-One loss.}
\usage{
calc_zero_one_loss(otp, on)
}
\arguments{
\item{otp}{Overall true positives}

\item{on}{Overall N}
}
\description{
Calculate Zero-One loss.
}
