% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary-calc-fmi.R
\name{FM_index_R}
\alias{FM_index_R}
\title{Calculate Fowlkes–Mallows index}
\usage{
FM_index_R(
  A1_clusters,
  A2_clusters,
  assume_sorted_vectors = FALSE,
  warn = dendextend_options("warn"),
  ...
)
}
\arguments{
\item{A1_clusters}{a numeric vector of cluster grouping (numeric) of items, with a name attribute of item name for each element from group A1. These are often obtained by using some k cut on a dendrogram.}

\item{A2_clusters}{a numeric vector of cluster grouping (numeric) of items, with a name attribute of item name for each element from group A2. These are often obtained by using some k cut on a dendrogram.}

\item{assume_sorted_vectors}{assume_sorted_vectors}

\item{warn}{warn}
}
\description{
The Fowlkes–Mallows index is the geometric mean of precision (PPV) and recall (TPR). It is generally used to compare the results of two clustering algorithms. It ranges from 0 to 1, with 1 indicating perfect classification.
}
\references{
Fowlkes, E. B.; Mallows, C. L. (1 September 1983). "A Method for Comparing Two Hierarchical Clusterings". Journal of the American Statistical Association. 78 (383): 553. doi:10.2307/2288117
dendextend implementation for multiclass - compare!
}
