% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiclass-calc-acc-macro.R,
%   R/multiclass-calc-aickin.R
\name{calc_acc_macro}
\alias{calc_acc_macro}
\alias{calc_acc_macro.default}
\alias{calc_acc_macro.table}
\alias{calc_acc_macro.data.frame}
\alias{calc_aickin.data.frame}
\title{Calculate Accuracy Macro.}
\usage{
calc_acc_macro(...)

\method{calc_acc_macro}{default}(tp, tn, n)

\method{calc_acc_macro}{table}(tbl)

\method{calc_acc_macro}{data.frame}(data, prediction, reference)

\method{calc_aickin}{data.frame}(
  data,
  prediction,
  reference,
  d = diag(1, nrow = nrow(tbl), ncol = ncol(tbl)),
  epsilon = 1e-07,
  ci.type = "aickin",
  ci.level = 0.95,
  maxiter = 1000
)
}
\arguments{
\item{tp}{Numeric vector of True Positives (TP) by class.}

\item{tn}{Numeric vector of True Negatives (TN) by class.}

\item{n}{Total number of observations.}

\item{tbl}{A table representing the input confusion matrix. This must always have prediction on rows and reference on columns, otherwise most functions in rmetrics will generate incorrect results.}

\item{data}{A data.frame containing the prediction and the reference.}

\item{prediction}{Character. The name of the variable in data that contains the predictions.}

\item{reference}{Character. The name of the variable in data that contains the reference values.}

\item{d}{The agreement matrix for tbl. Must have the same dimensions as tbl, with 1 to indicate agreeing scores, and 0 disagreeing scores. By default the diagonal is considered agreeing.}

\item{epsilon}{Convergence criterion. The algorithm stops when two consecutive \eqn{\alpha} estimates differ by less than epsilon. Default is 1e-7.}

\item{ci.type}{Either FALSE if no confidence intervals are desired or 'aickin' if the default CI's by Aickin should be calculated. If FALSE overwrites ci.level. Default is to calculate the CI.}

\item{ci.level}{A number between 0 and 1 for the levels of the confidence intervals that should be calculated.}

\item{maxiter}{Integer or double. Maximum number of iterations to try until convergence. Default is 1000.}
}
\description{
Calculate Accuracy Macro.
}
\section{Methods (by class)}{
\itemize{
\item \code{calc_acc_macro(default)}: 

\item \code{calc_acc_macro(table)}: 

\item \code{calc_acc_macro(data.frame)}: 

}}
\section{Functions}{
\itemize{
\item \code{calc_aickin(data.frame)}: 

}}
