% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiclass-calc-acc-macro.R
\name{calc_acc_macro}
\alias{calc_acc_macro}
\title{Calculate Accuracy macro (acc macro).}
\usage{
calc_acc_macro(tbl, tp, tn)
}
\arguments{
\item{tbl}{input confusion matrix of the form \code{table(prediction, reference)}}

\item{tp}{TP by class}

\item{tn}{TN by class}
}
\description{
Calculate Accuracy macro (acc macro).
}
