% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary-calc-af.R
\name{calc_af}
\alias{calc_af}
\alias{calc_af.default}
\alias{calc_af.table}
\alias{calc_af.data.frame}
\title{Calculate Adjusted F-score (af).}
\usage{
calc_af(...)

\method{calc_af}{default}(tp, fp, fn, tn, ...)

\method{calc_af}{table}(tbl, incr = FALSE, ...)

\method{calc_af}{data.frame}(data, prediction, reference, incr = FALSE, ...)
}
\arguments{
\item{...}{Additional arguments. Not used.}

\item{tp}{Numeric, True Positives (TP).}

\item{fp}{Numeric, False Positives (FP).}

\item{fn}{Numeric, Fase Negatives (FN).}

\item{tn}{Numeric, True Negatives (TN).}

\item{tbl}{A table representing the input confusion matrix. This must always have
prediction on rows and reference on columns, otherwise most functions in
rmetrics will generate incorrect results.}

\item{incr}{Double. Continuity correction to add to each cell of the table. FALSE
by default, which means the raw counts will be used.}

\item{data}{A data.frame containing the prediction and the reference.}

\item{prediction}{Character. The name of the variable in data that contains the predictions.}

\item{reference}{Character. The name of the variable in data that contains the reference
values.}
}
\value{
A numeric value that represents the Adjusted F-score (af) for the
binary classification model.

A numeric value that represents the Adjusted F-score (af) for the
binary classification model.
}
\description{
This function calculates the Adjusted F-score (af) which is a statistical
measure that combines both sensitivity and specificity of a binary
classification model, by taking into account both the false positive rate
(FPR) and the false negative rate (FNR) while penalizing more heavily the
FNR. This function uses the calc_f function to calculate the F2-score (F2)
and the inverse of the F0.5-score (inv05) and then applies the following
formula: sqrt(F2 * inv05).
}
\section{Methods (by class)}{
\itemize{
\item \code{calc_af(default)}: 

\item \code{calc_af(table)}: 

\item \code{calc_af(data.frame)}: 

}}
\section{Functions}{
\itemize{
\item \code{calc_af()}: 

}}
\examples{
calc_af(tp = 80, fp = 20, fn = 20, tn = 80)

}
\seealso{
\code{\link{calc_f}}
}
\concept{Classification Metrics}
\keyword{classification}
