% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary-calc-af.R
\name{calc_af}
\alias{calc_af}
\alias{calc_af.default}
\alias{calc_af.table}
\alias{calc_af.data.frame}
\title{Calculate Adjusted F-score (af).}
\usage{
calc_af(...)

\method{calc_af}{default}(tp, fp, fn, tn, ...)

\method{calc_af}{table}(tbl, ...)

\method{calc_af}{data.frame}(data, prediction, reference, ...)
}
\arguments{
\item{...}{Additional arguments. Not used.}

\item{tp}{Numeric, True Positives (TP).}

\item{fp}{Numeric, False Positives (FP).}

\item{fn}{Numeric, Fase Negatives (FN).}

\item{tn}{Numeric, True Negatives (TN).}

\item{tbl}{A table representing the input confusion matrix. This must always have prediction on rows and reference on columns, otherwise most functions in rmetrics will generate incorrect results.}

\item{data}{A data.frame containing the prediction and the reference.}

\item{prediction}{Character. The name of the variable in data that contains the predictions.}

\item{reference}{Character. The name of the variable in data that contains the reference values.}
}
\description{
Calculate Adjusted F-score (af).
}
\section{Methods (by class)}{
\itemize{
\item \code{calc_af(default)}: 

\item \code{calc_af(table)}: 

\item \code{calc_af(data.frame)}: 

}}
