% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiclass-calc-ari.R
\name{calc_ari}
\alias{calc_ari}
\alias{calc_ari.table}
\alias{calc_ari.data.frame}
\title{Calculate Adjusted Rand index.}
\usage{
calc_ari(...)

\method{calc_ari}{table}(tbl)

\method{calc_ari}{data.frame}(data, prediction, reference)
}
\arguments{
\item{tbl}{A table representing the input confusion matrix. This must always have prediction on rows and reference on columns, otherwise most functions in rmetrics will generate incorrect results.}

\item{data}{A data.frame containing the prediction and the reference.}

\item{prediction}{Character. The name of the variable in data that contains the predictions.}

\item{reference}{Character. The name of the variable in data that contains the reference values.}
}
\description{
Calculate Adjusted Rand index.
}
\section{Methods (by class)}{
\itemize{
\item \code{calc_ari(table)}: 

\item \code{calc_ari(data.frame)}: 

}}
