% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiclass-calc-aunp.R
\name{calc_aunp}
\alias{calc_aunp}
\title{Calculate overall AUC (aunp).}
\usage{
calc_aunp(data, classes, pos, n)
}
\arguments{
\item{data}{data with prediction, reference}

\item{classes}{confusion matrix classes}

\item{pos}{number of actual positives per class}

\item{n}{population or total number of samples per class}
}
\description{
Calculate overall AUC (aunp).
}
