% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary-calc-aupr.R
\name{calc_aupr}
\alias{calc_aupr}
\title{Calculate Area under the PR curve for each class (aupr).}
\usage{
calc_aupr(tp, fp, fn)
}
\arguments{
\item{tp}{Number of true positives in the contingency table.}

\item{fp}{Number of false positives in the contingency table.}

\item{fn}{Number of false negatives in the contingency table.}
}
\description{
Calculate Area under the PR curve for each class (aupr).
}
