% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary-calc-auroc.R
\name{calc_auroc}
\alias{calc_auroc}
\title{Calculate Area under the ROC curve for each class (auroc).}
\usage{
calc_auroc(tn, fp, tp, fn)
}
\arguments{
\item{tn}{Number of true negatives in the contingency table.}

\item{fp}{Number of false positives in the contingency table.}

\item{tp}{Number of true positives in the contingency table.}

\item{fn}{Number of false negatives in the contingency table.}
}
\description{
Calculate Area under the ROC curve for each class (auroc).
}
