% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary-calc-auroc.R
\name{calc_auroc}
\alias{calc_auroc}
\alias{calc_auroc.default}
\alias{calc_auroc.table}
\alias{calc_auroc.data.frame}
\title{Calculate Area under the ROC curve for each class (auroc).}
\usage{
calc_auroc(...)

\method{calc_auroc}{default}(tn, fp, tp, fn)

\method{calc_auroc}{table}(tbl)

\method{calc_auroc}{data.frame}(data, prediction, reference)
}
\arguments{
\item{tn}{Numeric, True Negatives (TN).}

\item{fp}{Numeric, False Positives (FP).}

\item{tp}{Numeric, True Positives (TP).}

\item{fn}{Numeric, Fase Negatives (FN).}

\item{tbl}{A table representing the input confusion matrix. This must always have prediction on rows and reference on columns, otherwise most functions in rmetrics will generate incorrect results.}

\item{data}{A data.frame containing the prediction and the reference.}

\item{prediction}{Character. The name of the variable in data that contains the predictions.}

\item{reference}{Character. The name of the variable in data that contains the reference values.}
}
\description{
Calculate Area under the ROC curve for each class (auroc).
}
\section{Methods (by class)}{
\itemize{
\item \code{calc_auroc(default)}: 

\item \code{calc_auroc(table)}: 

\item \code{calc_auroc(data.frame)}: 

}}
