% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary-calc-bbs.R
\name{calc_bbs}
\alias{calc_bbs}
\alias{calc_bbs.default}
\alias{calc_bbs.table}
\alias{calc_bbs.data.frame}
\title{Calculate Braun-Blanquet similarity.}
\usage{
calc_bbs(...)

\method{calc_bbs}{default}(tp, ppos, pos, ...)

\method{calc_bbs}{table}(tbl, incr = FALSE, ...)

\method{calc_bbs}{data.frame}(data, prediction, reference, incr = FALSE, ...)
}
\arguments{
\item{...}{Additional arguments. Not used.}

\item{tp}{Numeric, True Positives (TP).}

\item{ppos}{Number of positives in prediction vector (= TP + FP)}

\item{pos}{Number of positives in reference.}

\item{tbl}{A table representing the input confusion matrix. This must always have
prediction on rows and reference on columns, otherwise most functions in
rmetrics will generate incorrect results.}

\item{incr}{Double. Continuity correction to add to each cell of the table. FALSE
by default, which means the raw counts will be used.}

\item{data}{A data.frame containing the prediction and the reference.}

\item{prediction}{Character. The name of the variable in data that contains the predictions.}

\item{reference}{Character. The name of the variable in data that contains the reference
values.}
}
\description{
The Braun-Blanquet similarity is a measure of the degree to which two sets
of observations are similar. It is commonly used in ecology to compare
the composition of two different plant communities, but it can also be
applied to other types of data sets.
}
\details{
To calculate the Braun-Blanquet similarity, the following steps are followed:
\itemize{
\item For each observation in the first set of data, count the number of times it
occurs.
\item For each observation in the second set of data, count the number of times
it occurs.
\item For each observation that occurs in both sets of data, calculate the
minimum of the two counts and add this value to the Braun-Blanquet similarity
score.
\item Divide the Braun-Blanquet similarity score by the total number of
observations in both sets of data.
}

The resulting value is a measure of the degree to which the two sets of data
are similar, with values closer to 1 indicating greater similarity and
values closer to 0 indicating less similarity.
}
\section{Methods (by class)}{
\itemize{
\item \code{calc_bbs(default)}: 

\item \code{calc_bbs(table)}: 

\item \code{calc_bbs(data.frame)}: 

}}
\examples{
# Create some example data
set.seed(123)
reference <- factor(sample(c("A", "B", "C"), 100, replace = TRUE))
prediction <- factor(sample(c("A", "B", "C"), 100, replace = TRUE))
data <- data.frame(reference, prediction)

# Calculate Braun-Blanquet similarity using data frame
calc_bbs(data, prediction = "prediction", reference = "reference")

# Calculate Braun-Blanquet similarity using confusion matrix
tbl <- table(data)
calc_bbs(tbl)

# Calculate Braun-Blanquet similarity manually
tp <- sum(data$reference == "A" & data$prediction == "A")
ppos <- sum(data$prediction == "A")
pos <- sum(data$reference == "A")
calc_bbs(tp, ppos, pos)

}
