% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary-calc-bcd.R
\name{calc_bcd}
\alias{calc_bcd}
\alias{calc_bcd.default}
\alias{calc_bcd.table}
\alias{calc_bcd.data.frame}
\title{Calculate Bray-Curtis dissimilarity.}
\usage{
calc_bcd(...)

\method{calc_bcd}{default}(ppos, pos, n, ...)

\method{calc_bcd}{table}(tbl, ...)

\method{calc_bcd}{data.frame}(data, prediction, reference, ...)
}
\arguments{
\item{...}{Additional arguments. Not used.}

\item{ppos}{Number of positives in prediction vector (= TP + FP)}

\item{pos}{Number of positives in reference.}

\item{n}{Total number of observations.}

\item{tbl}{A table representing the input confusion matrix. This must always have
prediction on rows and reference on columns, otherwise most functions in
rmetrics will generate incorrect results.}

\item{data}{A data.frame containing the prediction and the reference.}

\item{prediction}{Character. The name of the variable in data that contains the predictions.}

\item{reference}{Character. The name of the variable in data that contains the reference
values.}
}
\description{
The Bray-Curtis dissimilarity is a measure of the degree to which two sets
of observations are different. It is commonly used in ecology to compare
the composition of two different plant communities, but it can also be
applied to other types of data sets.
}
\details{
To calculate the Bray-Curtis dissimilarity, the following steps are
followed:
\itemize{
\item For each observation in the first set of data, count the number of times
it occurs.
\item For each observation in the second set of data, count the number of times
it occurs.
\item For each observation that occurs in both sets of data, calculate the
absolute difference between the two counts and add this value to the Bray-
Curtis dissimilarity score.
\item Divide the Bray-Curtis dissimilarity score by the total number of
observations in both sets of data.
}

The resulting value is a measure of the degree to which the two sets of data
are different, with values closer to 1 indicating greater dissimilarity
and values closer to 0 indicating less dissimilarity.
}
\section{Methods (by class)}{
\itemize{
\item \code{calc_bcd(default)}: 

\item \code{calc_bcd(table)}: 

\item \code{calc_bcd(data.frame)}: 

}}
