% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary-calc-bcd.R
\name{calc_bcd}
\alias{calc_bcd}
\title{Calculate Bray-Curtis dissimilarity (bcd).}
\usage{
calc_bcd(ppos, pos, n)
}
\arguments{
\item{ppos}{Number of positives in predict vector}

\item{pos}{Number of actual positives}

\item{n}{Total number of samples.}
}
\description{
Calculate Bray-Curtis dissimilarity (bcd).
}
