% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiclass-calc-bennett-s.R
\name{calc_bennett_s}
\alias{calc_bennett_s}
\title{Calculate Bennett's S.}
\usage{
calc_bennett_s(tbl, otp, on)
}
\arguments{
\item{tbl}{input table}

\item{otp}{Overall true positives.}

\item{on}{Number of samples.}
}
\description{
Calculate Bennett's S.
}
