% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiclass-calc-cba.R
\name{calc_cba}
\alias{calc_cba}
\alias{calc_cba.default}
\alias{calc_cba.table}
\alias{calc_cba.data.frame}
\title{Calculate Class Balance Accuracy.}
\usage{
calc_cba(...)

\method{calc_cba}{default}(tp, fp, fn, ...)

\method{calc_cba}{table}(tbl, ...)

\method{calc_cba}{data.frame}(data, prediction, reference, ...)
}
\arguments{
\item{...}{Additional arguments. Not used.}

\item{tp}{Numeric vector of True Positives (TP) by class.}

\item{fp}{Numeric vector of False Positives (FP) by class.}

\item{fn}{Numeric vector of Fase Negatives (FN) by class.}

\item{tbl}{A table representing the input confusion matrix. This must always have prediction on rows and reference on columns, otherwise most functions in rmetrics will generate incorrect results.}

\item{data}{A data.frame containing the prediction and the reference.}

\item{prediction}{Character. The name of the variable in data that contains the predictions.}

\item{reference}{Character. The name of the variable in data that contains the reference values.}
}
\description{
Calculate Class Balance Accuracy.
}
\section{Methods (by class)}{
\itemize{
\item \code{calc_cba(default)}: 

\item \code{calc_cba(table)}: 

\item \code{calc_cba(data.frame)}: 

}}
