% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiclass-calc-conditional-entropy.R
\name{calc_conditional_entropy}
\alias{calc_conditional_entropy}
\alias{calc_conditional_entropy.table}
\alias{calc_conditional_entropy.data.frame}
\title{Calculate conditional entropy.}
\usage{
calc_conditional_entropy(...)

\method{calc_conditional_entropy}{table}(tbl, ...)

\method{calc_conditional_entropy}{data.frame}(
  data,
  prediction = "prediction",
  reference = "reference",
  ...
)
}
\arguments{
\item{...}{Additional arguments. Not used.

Computes the average conditional entropy between two vectors.}

\item{tbl}{A table representing the input confusion matrix. This must always have
prediction on rows and reference on columns, otherwise most functions in
rmetrics will generate incorrect results.}

\item{data}{A data.frame containing the prediction and the reference.}

\item{prediction}{Character. The name of the variable in data that contains the predictions.}

\item{reference}{Character. The name of the variable in data that contains the reference
values.}
}
\description{
Calculate conditional entropy.
}
\section{Methods (by class)}{
\itemize{
\item \code{calc_conditional_entropy(table)}: 

\item \code{calc_conditional_entropy(data.frame)}: 

}}
