% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiclass-calc-conditional-entropy.R
\name{calc_conditional_entropy}
\alias{calc_conditional_entropy}
\title{Calculate conditional entropy.}
\usage{
calc_conditional_entropy(tbl, pos)
}
\arguments{
\item{tbl}{confusion matrix}

\item{pos}{vector of actual positives per class}
}
\description{
Calculate conditional entropy.
}
