% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiclass-calc-cross-entropy.R
\name{calc_cross_entropy}
\alias{calc_cross_entropy}
\title{Calculate Cross entropy.}
\usage{
calc_cross_entropy(ppos, pos, n)
}
\arguments{
\item{ppos}{vector of positives in predict vector per class}

\item{pos}{vector of actual positives per class}

\item{n}{Total number of observations per class}
}
\description{
Calculate Cross entropy.
}
