% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary-calc-dind.R
\name{calc_dind}
\alias{calc_dind}
\alias{calc_dind.default}
\alias{calc_dind.table}
\alias{calc_dind.data.frame}
\title{Calculate Distance index (dind).}
\usage{
calc_dind(...)

\method{calc_dind}{default}(tn, fp, tp, fn, ...)

\method{calc_dind}{table}(tbl, ...)

\method{calc_dind}{data.frame}(data, prediction, reference, ...)
}
\arguments{
\item{...}{Additional arguments. Not used.}

\item{tn}{Numeric, True Negatives (TN).}

\item{fp}{Numeric, False Positives (FP).}

\item{tp}{Numeric, True Positives (TP).}

\item{fn}{Numeric, Fase Negatives (FN).}

\item{tbl}{A table representing the input confusion matrix. This must always have
prediction on rows and reference on columns, otherwise most functions in
rmetrics will generate incorrect results.}

\item{data}{A data.frame containing the prediction and the reference.}

\item{prediction}{Character. The name of the variable in data that contains the predictions.}

\item{reference}{Character. The name of the variable in data that contains the reference
values.}
}
\description{
Calculate Distance index (dind).
}
\section{Methods (by class)}{
\itemize{
\item \code{calc_dind(default)}: 

\item \code{calc_dind(table)}: 

\item \code{calc_dind(data.frame)}: 

}}
