% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary-calc-dor.R
\name{calc_dor}
\alias{calc_dor}
\alias{calc_dor.default}
\alias{calc_dor.table}
\alias{calc_dor.data.frame}
\title{Calculate Diagnostic odds ratio}
\usage{
calc_dor(...)

\method{calc_dor}{default}(tp, fn, tn, fp, ...)

\method{calc_dor}{table}(tbl, incr = FALSE, ...)

\method{calc_dor}{data.frame}(data, prediction, reference, incr = FALSE, ...)
}
\arguments{
\item{...}{Additional arguments. Not used.}

\item{tp}{Numeric, True Positives (TP).}

\item{fn}{Numeric, Fase Negatives (FN).}

\item{tn}{Numeric, True Negatives (TN).}

\item{fp}{Numeric, False Positives (FP).}

\item{tbl}{A table representing the input confusion matrix. This must always have
prediction on rows and reference on columns, otherwise most functions in
rmetrics will generate incorrect results.}

\item{incr}{Double. Continuity correction to add to each cell of the table. FALSE
by default, which means the raw counts will be used.}

\item{data}{A data.frame containing the prediction and the reference.}

\item{prediction}{Character. The name of the variable in data that contains the predictions.}

\item{reference}{Character. The name of the variable in data that contains the reference
values.}
}
\description{
The diagnostic odds ratio (DOR) is a metric that is used to evaluate
the performance of a diagnostic test. It is calculated as the ratio of
the odds of a positive test result in people with the disease to the
odds of a positive test result in people without the disease.
}
\details{
To calculate the DOR, the following formula is used:

DOR = (True positive rate / False positive rate) / (False negative rate /
True negative rate)

where the true positive rate is the ratio of true positive predictions to
the total number of actual positive samples, the false positive rate is
the ratio of false positive predictions to the total number of actual
negative samples, the false negative rate is the ratio of false
negative predictions to the total number of actual positive samples,
and the true negative rate is the ratio of true negative predictions to
the total number of actual negative samples.

The resulting value is a measure of the effectiveness of a diagnostic test,
with values closer to 1 indicating a weaker test and values closer to
infinity indicating a stronger test.
}
\section{Methods (by class)}{
\itemize{
\item \code{calc_dor(default)}: 

\item \code{calc_dor(table)}: 

\item \code{calc_dor(data.frame)}: 

}}
