% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary-calc-dp.R
\name{calc_dp}
\alias{calc_dp}
\title{Calculate Discriminant power (dp).}
\usage{
calc_dp(tn, fp, tp, fn)
}
\arguments{
\item{tn}{Number of true negatives in the contingency table.}

\item{fp}{Number of false positives in the contingency table.}

\item{tp}{Number of true positives in the contingency table.}

\item{fn}{Number of false negatives in the contingency table.}
}
\description{
Calculate Discriminant power (dp).
}
