% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary-calc-f05.R
\name{calc_f05}
\alias{calc_f05}
\alias{calc_f05.default}
\alias{calc_f05.table}
\alias{calc_f05.data.frame}
\title{Calculate F0.5 Score}
\usage{
calc_f05(...)

\method{calc_f05}{default}(tp, fp, fn)

\method{calc_f05}{table}(tbl)

\method{calc_f05}{data.frame}(data, prediction, reference)
}
\arguments{
\item{tp}{True Positives (TP).}

\item{fp}{False Positives (FP).}

\item{fn}{Fase Negatives (FN).}

\item{tbl}{A table representing the input confusion matrix. This must always have prediction on rows and reference on columns, otherwise most functions in rmetrics will generate incorrect results.}

\item{data}{A data.frame containing the prediction and the reference.}

\item{prediction}{Character. The name of the variable in data that contains the predictions.}

\item{reference}{Character. The name of the variable in data that contains the reference values.}
}
\description{
Calculate F0.5 Score
}
\section{Methods (by class)}{
\itemize{
\item \code{calc_f05(default)}: 

\item \code{calc_f05(table)}: 

\item \code{calc_f05(data.frame)}: 

}}
