% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary-calc-f1.R
\name{calc_f1}
\alias{calc_f1}
\alias{calc_f1.default}
\alias{calc_f1.table}
\alias{calc_f1.data.frame}
\title{Calculate F1 Score}
\usage{
calc_f1(...)

\method{calc_f1}{default}(tp, fp, fn)

\method{calc_f1}{table}(tbl)

\method{calc_f1}{data.frame}(data, prediction, reference)
}
\arguments{
\item{tp}{Numeric, True Positives (TP).}

\item{fp}{Numeric, False Positives (FP).}

\item{fn}{Numeric, Fase Negatives (FN).}

\item{tbl}{A table representing the input confusion matrix. This must always have prediction on rows and reference on columns, otherwise most functions in rmetrics will generate incorrect results.}

\item{data}{A data.frame containing the prediction and the reference.}

\item{prediction}{Character. The name of the variable in data that contains the predictions.}

\item{reference}{Character. The name of the variable in data that contains the reference values.}
}
\description{
Calculate F1 Score
}
\section{Methods (by class)}{
\itemize{
\item \code{calc_f1(default)}: 

\item \code{calc_f1(table)}: 

\item \code{calc_f1(data.frame)}: 

}}
