% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary-calc-fdr.R
\name{calc_fdr}
\alias{calc_fdr}
\alias{calc_fdr.default}
\alias{calc_fdr.table}
\alias{calc_fdr.data.frame}
\title{Calculate False Discovery Rate}
\usage{
calc_fdr(...)

\method{calc_fdr}{default}(fp, tp, ci.type, ci.level, ...)

\method{calc_fdr}{table}(tbl, ci.type, ci.level, ...)

\method{calc_fdr}{data.frame}(data, prediction, reference, ci.type, ci.level, ...)
}
\arguments{
\item{...}{Additional arguments. Not used.}

\item{fp}{Numeric, False Positives (FP).}

\item{tp}{Numeric, True Positives (TP).}

\item{ci.type}{Either FALSE if no confidence intervals are desired or one of "agresti.coull", "agresti-coull", "ac", "asymptotic", "normal", "wald", "clopper-pearson", "cp", "exact", "jeffreys", "bayes", and "wilson". If FALSE, overwrites ci.level.}

\item{ci.level}{A number between 0 and 1 for the levels of the confidence intervals that
should be calculated.}

\item{tbl}{A table representing the input confusion matrix. This must always have
prediction on rows and reference on columns, otherwise most functions in
rmetrics will generate incorrect results.}

\item{data}{A data.frame containing the prediction and the reference.}

\item{prediction}{Character. The name of the variable in data that contains the predictions.}

\item{reference}{Character. The name of the variable in data that contains the reference
values.}
}
\description{
The false discovery rate (FDR) is a measure of the proportion of false
positive results among all positive results in a statistical test. It
is calculated as the number of false positives divided by the total
number of positive results.
}
\details{
To calculate the FDR, the following formula is used:

FDR = (Number of false positives) / (Total number of positive results)

The resulting value is a proportion, with values closer to 0 indicating a
more reliable statistical test and values closer to 1 indicating a less
reliable test.
}
\section{Methods (by class)}{
\itemize{
\item \code{calc_fdr(default)}: 

\item \code{calc_fdr(table)}: 

\item \code{calc_fdr(data.frame)}: 

}}
