% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiclass-calc-fmi.R
\name{calc_fmi}
\alias{calc_fmi}
\title{Calculate Fowlkes–Mallows index}
\usage{
calc_fmi(otp, ofp, ofn)
}
\arguments{
\item{tp}{Number of true positives in the contingency table.}

\item{fp}{Number of false positives in the contingency table.}

\item{fn}{Number of false negatives in the contingency table.}
}
\description{
The Fowlkes–Mallows index is the geometric mean of precision (PPV) and recall (TPR). It is generally used to compare the results of two clustering algorithms. It ranges from 0 to 1, with 1 indicating perfect classification.
}
\references{
Fowlkes, E. B.; Mallows, C. L. (1 September 1983). "A Method for Comparing Two Hierarchical Clusterings". Journal of the American Statistical Association. 78 (383): 553. doi:10.2307/2288117
}
