% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiclass-calc-fnr-macro.R
\name{calc_fnr_macro}
\alias{calc_fnr_macro}
\title{Calculate FNR macro.}
\usage{
calc_fnr_macro(fn, tp)
}
\arguments{
\item{fn}{FN by class}

\item{tp}{TP by class}
}
\description{
Calculate FNR macro.
}
