% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiclass-calc-fnr-micro.R
\name{calc_fnr_micro}
\alias{calc_fnr_micro}
\alias{calc_fnr_micro.default}
\alias{calc_fnr_micro.table}
\alias{calc_fnr_micro.data.frame}
\title{Calculate FNR micro.}
\usage{
calc_fnr_micro(...)

\method{calc_fnr_micro}{default}(otp, n, ...)

\method{calc_fnr_micro}{table}(tbl, ...)

\method{calc_fnr_micro}{data.frame}(data, prediction, reference, ...)
}
\arguments{
\item{...}{Additional arguments. Not used.}

\item{otp}{Overall True Positives (OTP).}

\item{n}{Total number of observations.}

\item{tbl}{A table representing the input confusion matrix. This must always have prediction on rows and reference on columns, otherwise most functions in rmetrics will generate incorrect results.}

\item{data}{A data.frame containing the prediction and the reference.}

\item{prediction}{Character. The name of the variable in data that contains the predictions.}

\item{reference}{Character. The name of the variable in data that contains the reference values.}
}
\description{
Calculate FNR micro.
}
\section{Methods (by class)}{
\itemize{
\item \code{calc_fnr_micro(default)}: 

\item \code{calc_fnr_micro(table)}: 

\item \code{calc_fnr_micro(data.frame)}: 

}}
