% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary-calc-for.R
\name{calc_for}
\alias{calc_for}
\alias{calc_for.default}
\alias{calc_for.table}
\alias{calc_for.data.frame}
\title{Calculate False Omission Rate}
\usage{
calc_for(...)

\method{calc_for}{default}(fn, tn, ci.type, ci.level, ...)

\method{calc_for}{table}(tbl, ci.type, ci.level, ...)

\method{calc_for}{data.frame}(data, prediction, reference, ci.type, ci.level, ...)
}
\arguments{
\item{...}{Additional arguments. Not used.}

\item{fn}{Numeric, Fase Negatives (FN).}

\item{tn}{Numeric, True Negatives (TN).}

\item{ci.type}{Either FALSE if no confidence intervals are desired or one of "agresti.coull", "agresti-coull", "ac", "asymptotic", "normal", "wald", "clopper-pearson", "cp", "exact", "jeffreys", "bayes", and "wilson". If FALSE, overwrites ci.level.}

\item{ci.level}{A number between 0 and 1 for the levels of the confidence intervals that
should be calculated.}

\item{tbl}{A table representing the input confusion matrix. This must always have
prediction on rows and reference on columns, otherwise most functions in
rmetrics will generate incorrect results.}

\item{data}{A data.frame containing the prediction and the reference.}

\item{prediction}{Character. The name of the variable in data that contains the predictions.}

\item{reference}{Character. The name of the variable in data that contains the reference
values.}
}
\description{
The false omission rate (FOR) is a measure of the proportion of false
negative results among all negative results in a statistical test. It
is calculated as the number of false negatives divided by the total number
of negative results.
}
\details{
To calculate the FOR, the following formula is used:

FOR = (Number of false negatives) / (Total number of negative results)

The resulting value is a proportion, with values closer to 0 indicating a
more reliable statistical test and values closer to 1 indicating a less
reliable test.

The FOR is a useful metric to identify tests that are more likely
to accurately identify true negative results.
}
\section{Methods (by class)}{
\itemize{
\item \code{calc_for(default)}: 

\item \code{calc_for(table)}: 

\item \code{calc_for(data.frame)}: 

}}
