% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary-calc-gmean.R
\name{calc_gmean}
\alias{calc_gmean}
\alias{calc_gmean.default}
\alias{calc_gmean.table}
\alias{calc_gmean.data.frame}
\title{Calculate (adjusted) Geometric mean of TPR and TNR.}
\usage{
calc_gmean(...)

\method{calc_gmean}{default}(tn, fp, tp, fn, adjust = FALSE, ...)

\method{calc_gmean}{table}(tbl, adjust = FALSE, incr = FALSE, ...)

\method{calc_gmean}{data.frame}(data, prediction, reference, adjust = FALSE, incr = FALSE, ...)
}
\arguments{
\item{...}{Additional arguments. Not used.}

\item{tn}{Numeric, True Negatives (TN).}

\item{fp}{Numeric, False Positives (FP).}

\item{tp}{Numeric, True Positives (TP).}

\item{fn}{Numeric, Fase Negatives (FN).}

\item{adjust}{Should the GM be adjusted for the proportion of negatives?
FALSE by default.}

\item{tbl}{A table representing the input confusion matrix. This must always have
prediction on rows and reference on columns, otherwise most functions in
rmetrics will generate incorrect results.}

\item{incr}{Double. Continuity correction to add to each cell of the table. FALSE
by default, which means the raw counts will be used.}

\item{data}{A data.frame containing the prediction and the reference.}

\item{prediction}{Character. The name of the variable in data that contains the predictions.}

\item{reference}{Character. The name of the variable in data that contains the reference
values.}
}
\description{
The geometric mean is a measure of central tendency that is calculated by
taking the n-th root of the product of n numbers. In the context of binary
classification, the true positive rate (TPR) and true negative rate (TNR) are
two common metrics used to evaluate the performance of a model. The TPR is the
ratio of true positive predictions to the total number of positive samples,
and the TNR is the ratio of true negative predictions to the total number
of negative samples.
}
\details{
The geometric mean of TPR and TNR can be adjusted for the proportion of
negatives by setting \code{adjust = TRUE}.
}
\section{Methods (by class)}{
\itemize{
\item \code{calc_gmean(default)}: 

\item \code{calc_gmean(table)}: 

\item \code{calc_gmean(data.frame)}: 

}}
