% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary-calc-gmean.R
\name{calc_gmean}
\alias{calc_gmean}
\title{Calculate Geometric mean (gmean) of TPR and TNR.}
\usage{
calc_gmean(tn, fp, tp, fn)
}
\arguments{
\item{tn}{Number of true negatives in the contingency table.}

\item{fp}{Number of false positives in the contingency table.}

\item{tp}{Number of true positives in the contingency table.}

\item{fn}{Number of false negatives in the contingency table.}
}
\description{
Calculate Geometric mean (gmean) of TPR and TNR.
}
