% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiclass-calc-gwet-ac1.R
\name{calc_gwet_ac1}
\alias{calc_gwet_ac1}
\alias{calc_gwet_ac1.default}
\alias{calc_gwet_ac1.table}
\alias{calc_gwet_ac1.data.frame}
\title{Calculate Gwet AC1.}
\usage{
calc_gwet_ac1(...)

\method{calc_gwet_ac1}{default}(tp, fp, fn, n, ...)

\method{calc_gwet_ac1}{table}(tbl, ...)

\method{calc_gwet_ac1}{data.frame}(data, prediction, reference, ...)
}
\arguments{
\item{...}{Additional arguments. Not used.}

\item{tp}{Numeric vector of True Positives (TP) by class.}

\item{fp}{Numeric vector of False Positives (FP) by class.}

\item{fn}{Numeric vector of Fase Negatives (FN) by class.}

\item{n}{Total number of observations.}

\item{tbl}{A table representing the input confusion matrix. This must always have prediction on rows and reference on columns, otherwise most functions in rmetrics will generate incorrect results.}

\item{data}{A data.frame containing the prediction and the reference.}

\item{prediction}{Character. The name of the variable in data that contains the predictions.}

\item{reference}{Character. The name of the variable in data that contains the reference values.}
}
\description{
Calculate Gwet AC1.
}
\section{Methods (by class)}{
\itemize{
\item \code{calc_gwet_ac1(default)}: 

\item \code{calc_gwet_ac1(table)}: 

\item \code{calc_gwet_ac1(data.frame)}: 

}}
