% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary-calc-iba.R
\name{calc_iba}
\alias{calc_iba}
\alias{calc_iba.default}
\alias{calc_iba.table}
\alias{calc_iba.data.frame}
\title{Calculate Index of Balanced Accuracy.}
\usage{
calc_iba(...)

\method{calc_iba}{default}(tn, fp, tp, fn, alpha = 1, ...)

\method{calc_iba}{table}(tbl, alpha = 1, incr = FALSE, ...)

\method{calc_iba}{data.frame}(data, prediction, reference, alpha = 1, incr = FALSE, ...)
}
\arguments{
\item{...}{Additional arguments. Not used.}

\item{tn}{Numeric, True Negatives (TN).}

\item{fp}{Numeric, False Positives (FP).}

\item{tp}{Numeric, True Positives (TP).}

\item{fn}{Numeric, Fase Negatives (FN).}

\item{alpha}{Weight for TPR - TNR. By default 1.}

\item{tbl}{A table representing the input confusion matrix. This must always have
prediction on rows and reference on columns, otherwise most functions in
rmetrics will generate incorrect results.}

\item{incr}{Double. Continuity correction to add to each cell of the table. FALSE
by default, which means the raw counts will be used.}

\item{data}{A data.frame containing the prediction and the reference.}

\item{prediction}{Character. The name of the variable in data that contains the predictions.}

\item{reference}{Character. The name of the variable in data that contains the reference
values.}
}
\description{
Calculate Index of Balanced Accuracy.
}
\section{Methods (by class)}{
\itemize{
\item \code{calc_iba(default)}: 

\item \code{calc_iba(table)}: 

\item \code{calc_iba(data.frame)}: 

}}
