% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary-calc-informedness.R
\name{calc_informedness}
\alias{calc_informedness}
\alias{calc_informedness.default}
\alias{calc_informedness.table}
\alias{calc_informedness.data.frame}
\title{Calculate Informedness}
\usage{
calc_informedness(...)

\method{calc_informedness}{default}(tp, fn, tn, fp, ...)

\method{calc_informedness}{table}(tbl, ...)

\method{calc_informedness}{data.frame}(data, prediction, reference, ...)
}
\arguments{
\item{...}{Additional arguments. Not used.}

\item{tp}{Numeric, True Positives (TP).}

\item{fn}{Numeric, Fase Negatives (FN).}

\item{tn}{Numeric, True Negatives (TN).}

\item{fp}{Numeric, False Positives (FP).}

\item{tbl}{A table representing the input confusion matrix. This must always have prediction on rows and reference on columns, otherwise most functions in rmetrics will generate incorrect results.}

\item{data}{A data.frame containing the prediction and the reference.}

\item{prediction}{Character. The name of the variable in data that contains the predictions.}

\item{reference}{Character. The name of the variable in data that contains the reference values.}
}
\description{
Calculate Informedness
}
\section{Methods (by class)}{
\itemize{
\item \code{calc_informedness(default)}: 

\item \code{calc_informedness(table)}: 

\item \code{calc_informedness(data.frame)}: 

}}
