% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary-calc-is.R
\name{calc_is}
\alias{calc_is}
\alias{calc_is.default}
\alias{calc_is.table}
\alias{calc_is.data.frame}
\title{Calculate Information score.}
\usage{
calc_is(...)

\method{calc_is}{default}(tp, fp, fn, n, ...)

\method{calc_is}{table}(tbl, ...)

\method{calc_is}{data.frame}(data, prediction, reference, ...)
}
\arguments{
\item{...}{Additional arguments. Not used.}

\item{tp}{Numeric, True Positives (TP).}

\item{fp}{Numeric, False Positives (FP).}

\item{fn}{Numeric, Fase Negatives (FN).}

\item{n}{Total number of observations.}

\item{tbl}{A table representing the input confusion matrix. This must always have prediction on rows and reference on columns, otherwise most functions in rmetrics will generate incorrect results.}

\item{data}{A data.frame containing the prediction and the reference.}

\item{prediction}{Character. The name of the variable in data that contains the predictions.}

\item{reference}{Character. The name of the variable in data that contains the reference values.}
}
\description{
Calculate Information score.
}
\section{Methods (by class)}{
\itemize{
\item \code{calc_is(default)}: 

\item \code{calc_is(table)}: 

\item \code{calc_is(data.frame)}: 

}}
