% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiclass-calc-jaccard-overall.R
\name{calc_jaccard_overall}
\alias{calc_jaccard_overall}
\title{Calculate Mean overall Jaccard index.}
\usage{
calc_jaccard_overall(data)
}
\arguments{
\item{data}{data frame or matrix consisting of prediction and reference}
}
\description{
Calculate Mean overall Jaccard index.
}
