% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiclass-calc-joint-entropy.R
\name{calc_joint_entropy}
\alias{calc_joint_entropy}
\title{Calculate Joint entropy.}
\usage{
calc_joint_entropy(tbl)
}
\arguments{
\item{tbl}{input table}
}
\description{
Calculate Joint entropy.
}
