% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiclass-calc-kalpha.R
\name{calc_kalpha}
\alias{calc_kalpha}
\title{Calculate Unweighted Krippendorff's alpha (kalpha).}
\usage{
calc_kalpha(tbl, unbiased = TRUE)
}
\arguments{
\item{tbl}{The contingency table.}

\item{unbiased}{TRUE/FALSE. Should unbiased overall random accuracy be used?}
}
\description{
Calculate Unweighted Krippendorff's alpha (kalpha).
}
