% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiclass-calc-kalpha.R
\name{calc_kalpha}
\alias{calc_kalpha}
\title{Calculate Unweighted Krippendorff's alpha (kalpha).}
\usage{
calc_kalpha(tbl, otp)
}
\arguments{
\item{tbl}{The contingency table.}

\item{otp}{Overall true positives.}
}
\description{
Calculate Unweighted Krippendorff's alpha (kalpha).
}
