% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiclass-calc-kappa.R
\name{calc_kappa}
\alias{calc_kappa}
\title{Calculate Kappa, unbiased Kappa or Kappa no Prevalence.}
\usage{
calc_kappa(tbl, otp, on, unbiased = FALSE, prev = TRUE)
}
\arguments{
\item{tbl}{The contingency table.}

\item{otp}{Overall number of true positives in the contingency table.}

\item{on}{Overall number of observations}

\item{unbiased}{Logical, should 'normal' or unbiased overall random accuracy be used.}

\item{prev}{TRUE for Kappa and unbiased Kappa, FALSE for Kappa no prevalence.}
}
\description{
Calculate Kappa, unbiased Kappa or Kappa no Prevalence.
}
