% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary-calc-lift.R
\name{calc_lift}
\alias{calc_lift}
\title{Calculate Lift score (lift).}
\usage{
calc_lift(tp, fp, pos, neg)
}
\arguments{
\item{tp}{Number of true positives in the contingency table.}

\item{fp}{Number of false positives in the contingency table.}

\item{pos}{Number of actual positives}

\item{neg}{Number of actual negatives}
}
\description{
Calculate Lift score (lift).
}
