% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary-calc-markedness.R
\name{calc_markedness}
\alias{calc_markedness}
\alias{calc_markedness.default}
\alias{calc_markedness.table}
\alias{calc_markedness.data.frame}
\title{Calculate Markedness.}
\usage{
calc_markedness(...)

\method{calc_markedness}{default}(tp, fn, tn, fp)

\method{calc_markedness}{table}(tbl)

\method{calc_markedness}{data.frame}(data, prediction, reference)
}
\arguments{
\item{tp}{True Positives (TP).}

\item{fn}{Fase Negatives (FN).}

\item{tn}{True Negatives (TN).}

\item{fp}{False Positives (FP).}

\item{tbl}{A table representing the input confusion matrix. This must always have prediction on rows and reference on columns, otherwise most functions in rmetrics will generate incorrect results.}

\item{data}{A data.frame containing the prediction and the reference.}

\item{prediction}{Character. The name of the variable in data that contains the predictions.}

\item{reference}{Character. The name of the variable in data that contains the reference values.}
}
\description{
Calculate Markedness.
}
\section{Methods (by class)}{
\itemize{
\item \code{calc_markedness(default)}: 

\item \code{calc_markedness(table)}: 

\item \code{calc_markedness(data.frame)}: 

}}
