% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary-calc-mcc.R
\name{calc_mcc}
\alias{calc_mcc}
\alias{calc_mcc.default}
\alias{calc_mcc.table}
\alias{calc_mcc.data.frame}
\title{Calculate Matthews Correlation Coefficient}
\usage{
calc_mcc(...)

\method{calc_mcc}{default}(tp, tn, fp, fn, ...)

\method{calc_mcc}{table}(tbl, incr = FALSE, ...)

\method{calc_mcc}{data.frame}(data, prediction, reference, incr = FALSE, ...)
}
\arguments{
\item{...}{Additional arguments. Not used.}

\item{tp}{Numeric, True Positives (TP).}

\item{tn}{Numeric, True Negatives (TN).}

\item{fp}{Numeric, False Positives (FP).}

\item{fn}{Numeric, Fase Negatives (FN).}

\item{tbl}{A table representing the input confusion matrix. This must always have
prediction on rows and reference on columns, otherwise most functions in
rmetrics will generate incorrect results.}

\item{incr}{Double. Continuity correction to add to each cell of the table. FALSE
by default, which means the raw counts will be used.}

\item{data}{A data.frame containing the prediction and the reference.}

\item{prediction}{Character. The name of the variable in data that contains the predictions.}

\item{reference}{Character. The name of the variable in data that contains the reference
values.}
}
\description{
Calculate Matthews Correlation Coefficient
}
\section{Methods (by class)}{
\itemize{
\item \code{calc_mcc(default)}: 

\item \code{calc_mcc(table)}: 

\item \code{calc_mcc(data.frame)}: 

}}
