% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary-calc-mcc.R
\name{calc_mcc}
\alias{calc_mcc}
\alias{calc_mcc.default}
\alias{calc_mcc.table}
\alias{calc_mcc.data.frame}
\title{Calculate Matthews Correlation Coefficient}
\usage{
calc_mcc(...)

\method{calc_mcc}{default}(tp, tn, fp, fn)

\method{calc_mcc}{table}(tbl)

\method{calc_mcc}{data.frame}(data, prediction, reference)
}
\arguments{
\item{tp}{True Positives (TP).}

\item{tn}{True Negatives (TN).}

\item{fp}{False Positives (FP).}

\item{fn}{Fase Negatives (FN).}

\item{tbl}{A table representing the input confusion matrix. This must always have prediction on rows and reference on columns, otherwise most functions in rmetrics will generate incorrect results.}

\item{data}{A data.frame containing the prediction and the reference.}

\item{prediction}{Character. The name of the variable in data that contains the predictions.}

\item{reference}{Character. The name of the variable in data that contains the reference values.}
}
\description{
Calculate Matthews Correlation Coefficient
}
\section{Methods (by class)}{
\itemize{
\item \code{calc_mcc(default)}: 

\item \code{calc_mcc(table)}: 

\item \code{calc_mcc(data.frame)}: 

}}
