% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiclass-calc-mutual-information.R
\name{calc_mutual_information}
\alias{calc_mutual_information}
\title{Calculate Mutual information.}
\usage{
calc_mutual_information(tbl)
}
\arguments{
\item{tbl}{confusion matrix}
}
\description{
Calculate Mutual information.
}
