% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary-calc-net-benefit.R
\name{calc_net_benefit}
\alias{calc_net_benefit}
\alias{calc_net_benefit.default}
\alias{calc_net_benefit.table}
\alias{calc_net_benefit.data.frame}
\title{Calculate Net Benefit.}
\usage{
calc_net_benefit(...)

\method{calc_net_benefit}{default}(tp, fp, n, weight = 1, ...)

\method{calc_net_benefit}{table}(tbl, weight = 1, ...)

\method{calc_net_benefit}{data.frame}(data, prediction, reference, weight = 1, ...)
}
\arguments{
\item{...}{Additional arguments. Not used.}

\item{tp}{Numeric, True Positives (TP).}

\item{fp}{Numeric, False Positives (FP).}

\item{n}{Total number of observations.}

\item{weight}{The weight for FP in comparison to TP. By default 1.}

\item{tbl}{A table representing the input confusion matrix. This must always have
prediction on rows and reference on columns, otherwise most functions in
rmetrics will generate incorrect results.}

\item{data}{A data.frame containing the prediction and the reference.}

\item{prediction}{Character. The name of the variable in data that contains the predictions.}

\item{reference}{Character. The name of the variable in data that contains the reference
values.}
}
\description{
Calculate Net Benefit.
}
\section{Methods (by class)}{
\itemize{
\item \code{calc_net_benefit(default)}: 

\item \code{calc_net_benefit(table)}: 

\item \code{calc_net_benefit(data.frame)}: 

}}
