% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiclass-calc-oacc.R
\name{calc_oacc}
\alias{calc_oacc}
\alias{calc_oacc.default}
\alias{calc_oacc.table}
\alias{calc_oacc.data.frame}
\title{Calculate Calculate Overall Accuracy.}
\usage{
calc_oacc(...)

\method{calc_oacc}{default}(otp, n)

\method{calc_oacc}{table}(tbl)

\method{calc_oacc}{data.frame}(data, prediction, reference)
}
\arguments{
\item{otp}{Overall True Positives (OTP).}

\item{n}{Total number of observations.}

\item{tbl}{A table representing the input confusion matrix. This must always have prediction on rows and reference on columns, otherwise most functions in rmetrics will generate incorrect results.}

\item{data}{A data.frame containing the prediction and the reference.}

\item{prediction}{Character. The name of the variable in data that contains the predictions.}

\item{reference}{Character. The name of the variable in data that contains the reference values.}
}
\description{
The proportion of overall true positives, regardless of class. Identical to micro-averaging TPR.
}
\section{Methods (by class)}{
\itemize{
\item \code{calc_oacc(default)}: 

\item \code{calc_oacc(table)}: 

\item \code{calc_oacc(data.frame)}: 

}}
