% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary-calc-oc.R
\name{calc_oc}
\alias{calc_oc}
\title{Calculate Overlap coefficient (oc)}
\usage{
calc_oc(tp, ppos, pos)
}
\arguments{
\item{tp}{Number of true positives in the contingency table.}

\item{ppos}{Number of positives in predict vector}

\item{pos}{Number of actual positives}
}
\description{
Calculate Overlap coefficient (oc)
}
