% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary-calc-oc.R
\name{calc_oc}
\alias{calc_oc}
\alias{calc_oc.default}
\alias{calc_oc.table}
\alias{calc_oc.data.frame}
\title{Calculate Overlap Coefficient}
\usage{
calc_oc(...)

\method{calc_oc}{default}(tp, fp, fn, ...)

\method{calc_oc}{table}(tbl, incr = FALSE, ...)

\method{calc_oc}{data.frame}(data, prediction, reference, incr = FALSE, ...)
}
\arguments{
\item{...}{Additional arguments. Not used.}

\item{tp}{Numeric, True Positives (TP).}

\item{fp}{Numeric, False Positives (FP).}

\item{fn}{Numeric, Fase Negatives (FN).}

\item{tbl}{A table representing the input confusion matrix. This must always have
prediction on rows and reference on columns, otherwise most functions in
rmetrics will generate incorrect results.}

\item{incr}{Double. Continuity correction to add to each cell of the table. FALSE
by default, which means the raw counts will be used.}

\item{data}{A data.frame containing the prediction and the reference.}

\item{prediction}{Character. The name of the variable in data that contains the predictions.}

\item{reference}{Character. The name of the variable in data that contains the reference
values.}
}
\description{
Calculate Overlap Coefficient
}
\section{Methods (by class)}{
\itemize{
\item \code{calc_oc(default)}: 

\item \code{calc_oc(table)}: 

\item \code{calc_oc(data.frame)}: 

}}
