% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary-calc-op.R
\name{calc_op}
\alias{calc_op}
\alias{calc_op.default}
\alias{calc_op.table}
\alias{calc_op.data.frame}
\title{Calculate Optimized Precision.}
\usage{
calc_op(...)

\method{calc_op}{default}(tn, fp, tp, fn, ...)

\method{calc_op}{table}(tbl, ...)

\method{calc_op}{data.frame}(data, prediction, reference, ...)
}
\arguments{
\item{...}{Additional arguments. Not used.}

\item{tn}{Numeric, True Negatives (TN).}

\item{fp}{Numeric, False Positives (FP).}

\item{tp}{Numeric, True Positives (TP).}

\item{fn}{Numeric, Fase Negatives (FN).}

\item{tbl}{A table representing the input confusion matrix. This must always have prediction on rows and reference on columns, otherwise most functions in rmetrics will generate incorrect results.}

\item{data}{A data.frame containing the prediction and the reference.}

\item{prediction}{Character. The name of the variable in data that contains the predictions.}

\item{reference}{Character. The name of the variable in data that contains the reference values.}
}
\description{
Calculate Optimized Precision.
}
\section{Methods (by class)}{
\itemize{
\item \code{calc_op(default)}: 

\item \code{calc_op(table)}: 

\item \code{calc_op(data.frame)}: 

}}
