% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiclass-calc-oracc.R
\name{calc_oracc}
\alias{calc_oracc}
\title{Calculate Calculate overall random accuracy or unbiased overall random accuracy (oracc).}
\usage{
calc_oracc(tbl, unbiased = FALSE)
}
\arguments{
\item{tbl}{The contingency table.}

\item{unbiased}{TRUE/FALSE, should unbiased random accuracy be returned? FALSE by default.}
}
\description{
Calculate Calculate overall random accuracy or unbiased overall random accuracy (oracc).
}
