% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary-calc-phi.R
\name{calc_phi}
\alias{calc_phi}
\alias{calc_phi.default}
\alias{calc_phi.table}
\alias{calc_phi.data.frame}
\title{Calculate Phi Coefficient}
\usage{
calc_phi(...)

\method{calc_phi}{default}(tp, tn, fp, fn)

\method{calc_phi}{table}(tbl)

\method{calc_phi}{data.frame}(data, prediction, reference)
}
\arguments{
\item{tp}{True Positives (TP).}

\item{tn}{True Negatives (TN).}

\item{fp}{False Positives (FP).}

\item{fn}{Fase Negatives (FN).}

\item{tbl}{A table representing the input confusion matrix. This must always have prediction on rows and reference on columns, otherwise most functions in rmetrics will generate incorrect results.}

\item{data}{A data.frame containing the prediction and the reference.}

\item{prediction}{Character. The name of the variable in data that contains the predictions.}

\item{reference}{Character. The name of the variable in data that contains the reference values.}
}
\description{
Calculate Phi Coefficient
}
\section{Methods (by class)}{
\itemize{
\item \code{calc_phi(default)}: 

\item \code{calc_phi(table)}: 

\item \code{calc_phi(data.frame)}: 

}}
