% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary-calc-plr.R
\name{calc_plr}
\alias{calc_plr}
\alias{calc_plr.default}
\alias{calc_plr.table}
\alias{calc_plr.data.frame}
\title{Calculate Positive Likelihood Ratio}
\source{
Koopman, PAR (1984) Confidence intervals for the ratio of two binomial proportions. Biometrics; 513-517.
}
\usage{
calc_plr(...)

\method{calc_plr}{default}(tp, fn, fp, tn, ci.type, ci.level, ...)

\method{calc_plr}{table}(tbl, ci.type, ci.level, ...)

\method{calc_plr}{data.frame}(data, prediction, reference, ci.type, ci.level, ...)
}
\arguments{
\item{...}{Additional arguments. Not used.}

\item{tp}{Numeric, True Positives (TP).}

\item{fn}{Numeric, Fase Negatives (FN).}

\item{fp}{Numeric, False Positives (FP).}

\item{tn}{Numeric, True Negatives (TN).}

\item{ci.type}{Either FALSE if no confidence intervals are desired or 'koopman'. If FALSE overwrites ci.level.}

\item{ci.level}{A number between 0 and 1 for the levels of the confidence intervals that
should be calculated.}

\item{tbl}{A table representing the input confusion matrix. This must always have
prediction on rows and reference on columns, otherwise most functions in
rmetrics will generate incorrect results.}

\item{data}{A data.frame containing the prediction and the reference.}

\item{prediction}{Character. The name of the variable in data that contains the predictions.}

\item{reference}{Character. The name of the variable in data that contains the reference
values.}
}
\description{
Calculate Positive Likelihood Ratio
}
\section{Methods (by class)}{
\itemize{
\item \code{calc_plr(default)}: 

\item \code{calc_plr(table)}: 

\item \code{calc_plr(data.frame)}: 

}}
