% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary-calc-q.R
\name{calc_q}
\alias{calc_q}
\alias{calc_q.default}
\alias{calc_q.table}
\alias{calc_q.data.frame}
\title{Calculate Yule's Q.}
\usage{
calc_q(...)

\method{calc_q}{default}(tp, tn, fp, fn)

\method{calc_q}{table}(tbl)

\method{calc_q}{data.frame}(data, prediction, reference)
}
\arguments{
\item{tp}{Numeric, True Positives (TP).}

\item{tn}{Numeric, True Negatives (TN).}

\item{fp}{Numeric, False Positives (FP).}

\item{fn}{Numeric, Fase Negatives (FN).}

\item{tbl}{A table representing the input confusion matrix. This must always have prediction on rows and reference on columns, otherwise most functions in rmetrics will generate incorrect results.}

\item{data}{A data.frame containing the prediction and the reference.}

\item{prediction}{Character. The name of the variable in data that contains the predictions.}

\item{reference}{Character. The name of the variable in data that contains the reference values.}
}
\description{
Calculate Yule's Q.
}
\section{Methods (by class)}{
\itemize{
\item \code{calc_q(default)}: 

\item \code{calc_q(table)}: 

\item \code{calc_q(data.frame)}: 

}}
