% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary-calc-racc.R
\name{calc_racc}
\alias{calc_racc}
\alias{calc_racc.default}
\alias{calc_racc.table}
\alias{calc_racc.data.frame}
\title{Calculate (Unbiased) Random Accuracy}
\usage{
calc_racc(...)

\method{calc_racc}{default}(tp, fp, fn, n, unbiased = FALSE, ...)

\method{calc_racc}{table}(tbl, unbiased = FALSE, ...)

\method{calc_racc}{data.frame}(data, prediction, reference, unbiased = FALSE, ...)
}
\arguments{
\item{...}{Additional arguments. Not used.}

\item{tp}{Numeric, True Positives (TP).}

\item{fp}{Numeric, False Positives (FP).}

\item{fn}{Numeric, Fase Negatives (FN).}

\item{n}{Total number of observations.}

\item{unbiased}{Should unbiased random accuracy be calculated? FALSE by default.}

\item{tbl}{A table representing the input confusion matrix. This must always have
prediction on rows and reference on columns, otherwise most functions in
rmetrics will generate incorrect results.}

\item{data}{A data.frame containing the prediction and the reference.}

\item{prediction}{Character. The name of the variable in data that contains the predictions.}

\item{reference}{Character. The name of the variable in data that contains the reference
values.}
}
\description{
Calculate (Unbiased) Random Accuracy
}
\section{Methods (by class)}{
\itemize{
\item \code{calc_racc(default)}: 

\item \code{calc_racc(table)}: 

\item \code{calc_racc(data.frame)}: 

}}
