% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary-calc-raccu.R
\name{calc_raccu}
\alias{calc_raccu}
\title{Calculate Random acc Unbiased}
\usage{
calc_raccu(tp, fp, fn, n)
}
\arguments{
\item{tp}{Number of true positives}

\item{fp}{Number of false positives}

\item{fn}{Number of false negatives}

\item{n}{Total number of samples}
}
\description{
Calculate Random acc Unbiased
}
