% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiclass-calc-rci.R
\name{calc_rci}
\alias{calc_rci}
\alias{calc_rci.data.frame}
\alias{calc_rci.table}
\title{Calculate Relative classifier information (rci).}
\usage{
calc_rci(...)

\method{calc_rci}{data.frame}(data, prediction = "prediction", reference = "reference")

\method{calc_rci}{table}(tbl)
}
\arguments{
\item{data}{data}

\item{prediction}{prediciton}

\item{reference}{reference}

\item{tbl}{table}
}
\description{
Calculate Relative classifier information (rci).
}
\section{Methods (by class)}{
\itemize{
\item \code{calc_rci(data.frame)}: 

\item \code{calc_rci(table)}: 

}}
