% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiclass-calc-rci.R
\name{calc_rci}
\alias{calc_rci}
\title{Calculate Relative classifier information (rci).}
\usage{
calc_rci(tbl, pos, ppos, n)
}
\arguments{
\item{tbl}{confusion matrix}

\item{pos}{vector of actual positives per class}

\item{ppos}{vector of positives in predict vector per class}

\item{n}{Total number of observations}
}
\description{
Calculate Relative classifier information (rci).
}
