% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiclass-calc-response-entropy.R
\name{calc_response_entropy}
\alias{calc_response_entropy}
\title{Calculate Response Entropy.}
\usage{
calc_response_entropy(ppos, n)
}
\arguments{
\item{ppos}{vector of positives in predict vector per class}

\item{n}{Total number of observations}
}
\description{
Calculate Response Entropy.
}
