% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiclass-calc-tnr-micro.R
\name{calc_tnr_micro}
\alias{calc_tnr_micro}
\alias{calc_tnr_micro.default}
\alias{calc_tnr_micro.table}
\alias{calc_tnr_micro.data.frame}
\title{Calculate TNR Micro.}
\usage{
calc_tnr_micro(...)

\method{calc_tnr_micro}{default}(tn, fp, ...)

\method{calc_tnr_micro}{table}(tbl, ...)

\method{calc_tnr_micro}{data.frame}(data, prediction, reference, ...)
}
\arguments{
\item{...}{Additional arguments. Not used.}

\item{tn}{Numeric vector of True Negatives (TN) by class.}

\item{fp}{Numeric vector of False Positives (FP) by class.}

\item{tbl}{A table representing the input confusion matrix. This must always have prediction on rows and reference on columns, otherwise most functions in rmetrics will generate incorrect results.}

\item{data}{A data.frame containing the prediction and the reference.}

\item{prediction}{Character. The name of the variable in data that contains the predictions.}

\item{reference}{Character. The name of the variable in data that contains the reference values.}
}
\description{
Calculate TNR Micro.
}
\section{Methods (by class)}{
\itemize{
\item \code{calc_tnr_micro(default)}: 

\item \code{calc_tnr_micro(table)}: 

\item \code{calc_tnr_micro(data.frame)}: 

}}
