% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiclass-calc-tpr-macro.R
\name{calc_tpr_macro}
\alias{calc_tpr_macro}
\title{Calculate TPR macro.}
\usage{
calc_tpr_macro(tp, fn)
}
\arguments{
\item{tp}{TP by class}

\item{fn}{FN by class}
}
\description{
Calculate TPR macro.
}
